\name{Accident}
\docType{data}
\alias{Accident}
\title{Ship Accidents }
\description{
  a cross-section

 \emph{number of observations} :  40

}
\usage{data(Accident)}
\format{A dataframe containing :
  \describe{
    \item{type}{ship type, a factor with levels (A,B,C,D,E)}
    \item{constr}{year constructed, a factor with levels (C6064,C6569,C7074,C7579)}
    \item{operate}{year operated, a factor with levels (O6074,O7579)}
    \item{months}{measure of service amount}
    \item{acc}{accidents}
  }
}
\source{
    McCullagh, P.  and  J.  Nelder (1983) \emph{Generalized Linear Models}, New York:Chapman and Hall.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}, Table F21.3.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
