\name{Airline}
\docType{data}
\alias{Airline}
\title{Cost for U.S. Airlines }
\description{
  a panel of 6 observations  from 1970  to 1984

 \emph{number of observations} :  90

 \emph{observation} :  production units

   \emph{country} :  United States

}
\usage{data(Airline)}
\format{A dataframe containing :
  \describe{
    \item{airline}{airline}
    \item{year}{year}
    \item{cost}{total cost, in \$1,000}
    \item{output}{output, in revenue passenger miles, 
      index number}
    \item{pf}{fuel price}
    \item{lf}{load factor, the average capacity utilization of the fleet}
  }
}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}, Table F7.1.

}
\seealso{
  \code{\link{Index.Source}}, \code{\link{Index.Economics}},
  \code{\link{Index.Econometrics}}, 
  \code{\link{Index.Observations}},
  \code{\link{Index.Time.Series}}
}
\keyword{datasets}
