\name{Airq}
\docType{data}
\alias{Airq}
\title{Air Quality for Californian Metropolitan Areas }
\description{
  a cross-section from 1972

 \emph{number of observations} :  30

 \emph{observation} :  regional

   \emph{country} :  United States

}
\usage{data(Airq)}
\format{A dataframe containing :
  \describe{
    \item{airq}{indicator of air quality (the lower the better)}
    \item{vala}{value added of companies (in thousands of dollars)}
    \item{rain}{amount of rain (in inches)}
    \item{coas}{is it a coastal area ? }
    \item{dens}{population density (per square mile)}
    \item{medi}{average income per head (in US dollars)}
  }
}
\references{
    Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{https://feb.kuleuven.be/GME/}, chapter 4.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}

\keyword{datasets}
