\name{BudgetFood}
\docType{data}
\alias{BudgetFood}
\title{Budget Share of Food for Spanish Households }
\description{
  a cross-section from 1980

 \emph{number of observations} :  23972

 \emph{observation} :  households

   \emph{country} :  Spain

}
\usage{data(BudgetFood)}
\format{A dataframe containing :
  \describe{
    \item{wfood}{percentage of total expenditure which the household has spent on food}
    \item{totexp}{total expenditure of the household}
    \item{age}{age of reference person in the household}
    \item{size}{size of the household}
    \item{town}{
      size of the town where the household is 
      placed  categorized into 5 groups: 1 for 
      small towns, 5 for big ones
    }
    \item{sex}{sex of reference person (man,woman)}
  }
}
\source{
    Delgado, A.  and  Juan  Mora (1998) \dQuote{Testing non--nested semiparametric models : an application to Engel curves specification}, \emph{Journal of Applied Econometrics}, \bold{13(2)}, 145--162.

}
\references{
      Journal of Applied Econometrics data archive : \url{http://qed.econ.queensu.ca/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
