\name{BudgetItaly}
\docType{data}
\alias{BudgetItaly}
\title{Budget Shares for Italian Households }
\description{
  a cross-section from 1973 to 1992

 \emph{number of observations} :  1729

 \emph{observation} :  households

   \emph{country} :  Italy

}
\usage{data(BudgetItaly)}
\format{A dataframe containing :
  \describe{
    \item{wfood}{food share}
    \item{whouse}{housing and fuels share}
    \item{wmisc}{miscellaneous share}
    \item{pfood}{food price}
    \item{phouse}{housing and fuels price}
    \item{pmisc}{miscellaneous price}
    \item{totexp}{total expenditure}
    \item{year}{year }
    \item{income}{income}
    \item{size}{household size}
    \item{pct}{cellule weight}
  }
}
\source{
    Bollino, Carlo Andrea,  Frederico   Perali and  Nicola  Rossi (2000) \dQuote{Linear household technologies}, \emph{Journal of Applied Econometrics}, \bold{15(3)}, 253--274.

}
\references{
      Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}

\keyword{datasets}
