\name{BudgetUK}
\docType{data}
\alias{BudgetUK}
\title{Budget Shares of British Households }
\description{
  a cross-section from 1980 to 1982

 \emph{number of observations} :  1519

 \emph{observation} :  households

   \emph{country} :  United Kingdom

}
\usage{data(BudgetUK)}
\format{A dataframe containing :
  \describe{
    \item{wfood}{budget share for food expenditure}
    \item{wfuel}{budget share for fuel expenditure}
    \item{wcloth}{
      budget share for clothing expenditure
    }
    \item{walc}{
      budget share for alcohol expenditure
    }
    \item{wtrans}{
      budget share for transport expenditure
    }
    \item{wother}{
      budget share for other good expenditure
    }
    \item{totexp}{
      total household expenditure (rounded to 
      the nearest 10 UK pounds sterling)
    }
    \item{income}{
      total net household income (rounded to the
      nearest 10 UK pounds sterling)
    }
    \item{age}{age of household head}
    \item{children}{number of children}
  }
}
\source{
    Blundell, Richard, Alan Duncan and Krishna 
    Pendakur (1998) \dQuote{Semiparametric 
    estimation and consumer demand}, \emph{Journal 
    of Applied Econometrics}, \bold{13(5)}, 435--462.
}
\references{
      Journal of Applied Econometrics data archive : \url{http://qed.econ.queensu.ca/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
