\name{Bwages}
\docType{data}
\alias{Bwages}
\title{Wages in Belgium }
\description{
  a cross-section from 1994

 \emph{number of observations} :  1472

 \emph{observation} :  individuals

   \emph{country} :  Belgium

}
\usage{data(Bwages)}
\format{A dataframe containing :
  \describe{
    \item{wage}{ gross hourly wage rate in euro}
    \item{educ}{ education level from 1 [low] to 5 [high]}
    \item{exper}{years of experience}
    \item{sex}{a factor with levels (males,female)}
  }
}
\source{
    European Community Household Panel.

}
\references{
    Verbeek, Marno (2004) \emph{A Guide to Modern Econometrics}, John Wiley and Sons, chapter 3.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}

\keyword{datasets}
