\name{Capm}
\docType{data}
\alias{Capm}
\title{Stock Market Data }
\description{
  monthly observations from 1960--01 to 2002--12

 \emph{number of observations} :  516

}
\usage{data(Capm)}
\format{A time serie containing :
  \describe{
    \item{rfood}{excess returns food industry}
    \item{rdur}{excess returns durables industry}
    \item{rcon}{excess returns construction industry}
    \item{rmrf}{excess returns market portfolio}
    \item{rf}{riskfree return}
  }
}
\source{
    most of the above data are from Kenneth French's data library
    at \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}.

}
\references{
    Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{https://feb.kuleuven.be/GME/}, chapter 2.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}

\keyword{datasets}
