\name{Cigarette}
\docType{data}
\alias{Cigarette}
\title{The Cigarette Consumption Panel Data Set }
\description{
  a panel of 48 observations  from 1985  to 1995

 \emph{number of observations} :  528

 \emph{observation} :  regional

   \emph{country} :  United States

}
\usage{data(Cigarette)}
\format{A dataframe containing :
  \describe{
    \item{state}{state}
    \item{year}{year}
    \item{cpi}{consumer price index}
    \item{pop}{state population}
    \item{packpc}{number of packs per capita}
    \item{income}{state personal income (total, nominal)}
    \item{tax}{average state, federal, and average local excise taxes for fiscal year}
    \item{avgprs}{average price during fiscal year, including sales taxes}
    \item{taxs}{average excise taxes for fiscal year, including sales taxes}
  }
}
\source{
    Professor Jonathan Gruber, MIT.

}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers, chapter 10.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
