\name{Clothing}
\docType{data}
\alias{Clothing}
\title{Sales Data of Men's Fashion Stores }
\description{
  a cross-section from 1990

 \emph{number of observations} :  400

 \emph{observation} :  production units

   \emph{country} :  Netherland

}
\usage{data(Clothing)}
\format{A dataframe containing :
  \describe{
    \item{tsales}{annual sales in Dutch guilders }
    \item{sales}{sales per square meter}
    \item{margin}{gross-profit-margin}
    \item{nown}{number of owners (managers)}
    \item{nfull}{number of full-timers}
    \item{npart}{number of part-timers}
    \item{naux}{number of helpers (temporary workers)}
    \item{hoursw}{total number of hours worked}
    \item{hourspw}{number of hours worked per worker}
    \item{inv1}{investment in shop-premises}
    \item{inv2}{investment in automation.}
    \item{ssize}{sales floorspace of the store (in m$^2$).}
    \item{start}{year start of business}
  }
}
\references{
    Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{http://www.econ.kuleuven.ac.be/GME}, chapter 3.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}

\keyword{datasets}
