\name{Computers}
\docType{data}
\alias{Computers}
\title{Prices of Personal Computers }
\description{
  a cross-section from 1993 to 1995

 \emph{number of observations} :  6259

 \emph{observation} :  goods

   \emph{country} :  United States

}
\usage{data(Computers)}
\format{A dataframe containing :
  \describe{
    \item{price}{price in US dollars of 486 PCs}
    \item{speed}{clock speed in MHz}
    \item{hd}{size of hard drive in MB}
    \item{ram}{size of Ram in in MB}
    \item{screen}{size of screen in inches }
    \item{cd}{is a CD-ROM present  ?}
    \item{multi}{is a multimedia kit (speakers, sound card) included ?}
    \item{premium}{is the manufacturer was a "premium" firm (IBM, COMPAQ) ?}
    \item{ads}{number of 486 price listings for each month}
    \item{trend}{time trend indicating month starting from January of 1993 to November of 1995.}
  }
}
\source{
    Stengos, T.  and  E.  Zacharias (2005) \dQuote{Intertemporal pricing and price discrimination : a semiparametric hedonic analysis of the personal computer market}, \emph{Journal of Applied Econometrics}, \bold{forthcoming}.

}
\references{
      Journal of Applied Econometrics data archive : \url{http://qed.econ.queensu.ca/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
