\name{Consumption}
\docType{data}
\alias{Consumption}
\title{Quarterly Data on Consumption and Expenditure }
\description{
  quarterly observations from 1947-1 to 1996-4

 \emph{number of observations} :  200

 \emph{observation} :  country

   \emph{country} :  Canada

}
\usage{data(Consumption)}
\format{A time series containing :
  \describe{
    \item{yd}{ personal disposable income, 1986 dollars}
    \item{ce}{personal consumption expenditure, 1986 dollars}
  }
}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 1, 3, 4, 6, 9, 10, 14 and 15.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}

\keyword{datasets}
