\name{CPSch3}
\docType{data}
\alias{CPSch3}
\title{Earnings from the Current Population Survey }
\description{
  a cross-section from 1998

 \emph{number of observations} :  11130

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(CPSch3)}
\format{A dataframe containing :
  \describe{
    \item{year}{survey year}
    \item{ahe}{average hourly earnings}
    \item{sex}{a factor with levels (male,female)}
  }
}
\source{
    Bureau of labor statistics, U.S. Department of Labor \url{http://www.bls.gov}.

}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers, chapter 3.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}

\keyword{datasets}
