\name{Diamond}
\docType{data}
\alias{Diamond}
\title{Pricing the C's of Diamond Stones }
\description{
  a cross-section from 2000

 \emph{number of observations} :  308

 \emph{observation} :  goods

   \emph{country} :  Singapore

}
\usage{data(Diamond)}
\format{A dataframe containing :
  \describe{
    \item{carat}{weight of diamond stones in carat unit}
    \item{colour}{a factor with levels (D,E,F,G,H,I)}
    \item{clarity}{a factor with levels (IF,VVS1,VVS2,VS1,VS2)}
    \item{certification}{certification body, a factor with levels (GIA,IGI,HRD)}
    \item{price}{price in Singapore \$}
  }
}
\source{
    Chu, Singfat (2001) \dQuote{Pricing the C's of Diamond Stones}, \emph{Journal of Statistics Education}, \bold{9(2)}.

}
\references{
      Journal of Statistics Education's data archive : 
  \url{http://www.amstat.org/publications/jse/jse_data_archive.htm}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
