\name{Doctor}
\docType{data}
\alias{Doctor}
\title{Number of Doctor Visits }
\description{
  a cross-section from 1986

 \emph{number of observations} :  485

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Doctor)}
\format{A dataframe containing :
  \describe{
    \item{doctor}{the number of doctor visits}
    \item{children}{the number of children in the household}
    \item{access}{is a measure of access to health care}
    \item{health}{a measure of health status (larger positive numbers are  associated with poorer health)}
  }
}
\source{
    Gurmu, Shiferaw (1997) \dQuote{Semiparametric estimation of hurdle regression models with an application to medicaid utilization}, \emph{Journal of Applied Econometrics}, \bold{12(3)}, 225-242.

}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, \url{http://www.econ.queensu.ca/ETM/}, chapter 11.

      Journal of Applied Econometrics data archive : \url{http://qed.econ.queensu.ca/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
