\name{Earnings}
\docType{data}
\alias{Earnings}
\title{Earnings for Three Age Groups }
\description{
  a cross-section from 1988-1989

 \emph{number of observations} :  4266

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Earnings)}
\format{A dataframe containing :
  \describe{
    \item{age}{age groups, a factor with levels (g1,g2,g3)}
    \item{y}{average annual earnings, in 1982 US dollars}
  }
}
\source{
    Mills, Jeffery A.  and  Sourushe  Zandvakili (1997) \dQuote{Statistical Inference via  Bootstrapping for Measures of Inequality}, \emph{Journal of Applied Econometrics}, \bold{12(2)}, pp. 133-150.

}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, \url{http://www.econ.queensu.ca/ETM/}, chapter 5 and 7.

      Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}

\keyword{datasets}
