\name{Electricity}
\docType{data}
\alias{Electricity}
\title{Cost Function for Electricity Producers }
\description{
  a cross-section from 1970 to 1970

 \emph{number of observations} :  158

 \emph{observation} :  production units

   \emph{country} :  United States

}
\usage{data(Electricity)}
\format{A dataframe containing :
  \describe{
    \item{cost}{total cost}
    \item{q}{total output}
    \item{pl}{wage rate}
    \item{sl}{cost share for labor}
    \item{pk}{capital price index}
    \item{sk}{cost share for capital}
    \item{pf}{fuel price}
    \item{sf}{cost share for fuel}
  }
}
\source{
    Christensen, L.  and  W. H.  Greene (1976) \dQuote{Economies of scale in U.S. electric power generation}, \emph{Journal of Political Economy}, \bold{84}, 655-676.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}, chapter 4, 317-320.

    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, \url{http://fhayashi.fc2web.com/hayashi_econometrics.htm}, chapter 1, 76-84.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}

\keyword{datasets}
