\name{Fatality}
\docType{data}
\alias{Fatality}
\title{Drunk Driving Laws and Traffic Deaths }
\description{
  a panel of 48 observations  from 1982  to 1988

 \emph{number of observations} :  336

 \emph{observation} :  regional

   \emph{country} :  United States

}
\usage{data(Fatality)}
\format{A dataframe containing :
  \describe{
    \item{state}{state ID code}
    \item{year}{year}
    \item{mrall}{traffic fatality rate (deaths per 10000)}
    \item{beertax}{tax on case of beer}
    \item{mlda}{minimum legal drinking age}
    \item{jaild}{mandatory jail sentence ?}
    \item{comserd}{mandatory community service ?}
    \item{vmiles}{average miles per driver}
    \item{unrate}{unemployment rate}
    \item{perinc}{per capita personal income}
  }
}
\source{
    Pr. Christopher J. Ruhm, Department of Economics, University of North Carolina.

}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers, \url{http://wps.aw.com/aw_stockwatsn_economtrcs_1}, chapter 8.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
