\name{Forward}
\docType{data}
\alias{Forward}
\title{Exchange Rates of US Dollar Against Other Currencies }
\description{
  monthly observations from 1979--01 to 2001--12

 \emph{number of observations} :  276

}
\usage{data(Forward)}
\format{A time series containing :
  \describe{
    \item{usdbp}{ exchange rate USD/British Pound Sterling}
    \item{usdeuro}{ exchange rate US D/Euro }
    \item{eurobp}{ exchange rate Euro/Pound}
    \item{usdbp1}{1 month forward rate USD/Pound}
    \item{usdeuro1}{1 month forward rate USD/Euro}
    \item{eurobp1}{1 month forward rate Euro/Pound}
    \item{usdbp3}{3 month forward rate USD/Pound}
    \item{usdeuro3}{month forward rate USD/Euro}
    \item{eurobp3}{month forward rate Euro/Pound}
  }
}
\source{
    Datastream                                                                   .

}
\references{
    Verbeek, Marno (2004) \emph{A Guide to Modern Econometrics}, John Wiley and Sons, chapter 4.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
