\name{Forward2}
\docType{data}
\alias{Forward2}
\title{Exchange Rates of Us Dollar Against Other Currencies }
\description{
  monthly observations from 1979--01 to 2001--12

 \emph{number of observations} :  276

}
\usage{data(Forward2)}
\format{A time serie containing :
  \describe{
    \item{usdbp}{ exchange rate USDollar/British Pound Sterling}
    \item{usdeuro}{ exchange rate USDollar/euro }
    \item{eurobp}{ exchange rate euro/pound}
    \item{usdbp1}{1 month forward rate USD/pound}
    \item{usdeuro1}{1 month forward rate USD/euro}
    \item{eurobp1}{1 month forward rate euro/pound}
    \item{usdbp3}{3 month forward rate USD/pound}
    \item{usdeuro3}{month forward rate USD/euro}
    \item{eurobp3}{month forward rate euro/pound}
  }
}
\source{
    Datastream                                                                   .

}
\references{
    Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{http://www.econ.kuleuven.ac.be/GME}, chapter 4.

}
\seealso{\code{\link{Index.Books}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}, \code{\link{Index.Time.Series}}}

\keyword{datasets}
