\name{Griliches}
\docType{data}
\alias{Griliches}
\title{Wage Data }
\description{
  a cross-section from 1980

 \emph{number of observations} :  758

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Griliches)}
\format{A dataframe containing :
  \describe{
    \item{rns}{residency in the southern states (first observation) ?}
    \item{rns80}{same variable for 1980}
    \item{mrt}{married  (first observation) ?}
    \item{mrt80}{same variable for 1980}
    \item{smsa}{residency in metropolitan areas (first observation) ?}
    \item{smsa80}{same variable for 1980}
    \item{med}{mother's education in years}
    \item{iq}{IQ score}
    \item{kww}{score on the ``knowledge of the world of work'' test}
    \item{year}{year of the observation}
    \item{age}{age (first observation)}
    \item{age80}{same variable for 1980}
    \item{school}{completed years of schooling (first observation)}
    \item{school80}{same variable for 1980}
    \item{expr}{experience in years (first observation)}
    \item{expr80}{same variable for 1980}
    \item{tenure}{tenure in years (first observation)}
    \item{tenure80}{same variable for 1980}
    \item{lw}{log wage (first observation)}
    \item{lw80}{same variable for 1980}
  }
}
\source{
    Blackburn, M.  and   Neumark D. (1992) \dQuote{Unobserved ability, efficiency wages, and interindustry wage differentials}, \emph{Quarterly Journal of Economics}, \bold{107}, 1421-1436.

}
\references{
    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, \url{http://fhayashi.fc2web.com/}, chapter 3, 250-256.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
