\name{HI}
\docType{data}
\alias{HI}
\title{Health Insurance and Hours Worked By Wives }
\description{
  a cross-section from 1993

 \emph{number of observations} :  22272

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(HI)}
\format{A dataframe containing :
  \describe{
    \item{whrswk}{hours worked per week by wife}
    \item{hhi}{wife covered by husband's HI ?}
    \item{whi}{wife has HI thru her job ?}
    \item{hhi2}{husband has HI thru own job ?}
    \item{education}{a factor with levels, "<9years", "9-11years", "12years", "13-15years", "16years", ">16years"}
    \item{race}{one of white, black, other}
    \item{hispanic}{hispanic ?}
    \item{experience}{years of potential work experience}
    \item{kidslt6}{number of kids under age of 6}
    \item{kids618}{number of kids 6--18 years old}
    \item{husby}{husband's income in thousands of dollars}
    \item{region}{one of other, northcentral, south, west}
    \item{wght}{sampling weight}
  }
}
\source{
    Olson, Craig A. (1998) \dQuote{A comparison of parametric and semiparametric estimates of the effect of spousal healpth insurance coverage on weekly hours worked by wiwes}, \emph{Journal of Applied Econometrics}, \bold{13(5)}, september--october, 543--565.

}
\references{
      Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
