\name{Housing}
\docType{data}
\alias{Housing}
\title{Sales Prices of Houses in the City of Windsor }
\description{
  a cross-section from 1987

 \emph{number of observations} :  546

 \emph{observation} :  goods

   \emph{country} :  Canada

}
\usage{data(Housing)}
\format{A dataframe containing :
  \describe{
    \item{price}{sale price of a house}
    \item{lotsize}{the lot size of a property in square feet}
    \item{bedrooms}{number of bedrooms}
    \item{bathrms}{number of full bathrooms}
    \item{stories}{number of stories excluding basement}
    \item{driveway}{does the house has a driveway ?}
    \item{recroom}{does the house has a recreational room ?}
    \item{fullbase}{does the house has a full finished basement ?}
    \item{gashw}{does the house uses gas for hot water heating ?}
    \item{airco}{does the house has central air conditioning ?}
    \item{garagepl}{number of garage places}
    \item{prefarea}{is the house located in the preferred neighbourhood of the city ?}
  }
}
\source{
    Anglin, P.M.  and  R.  Gencay (1996) \dQuote{Semiparametric estimation of a hedonic price function}, \emph{Journal of Applied Econometrics}, \bold{11(6)}, 633-648.

}
\references{
    Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{https://feb.kuleuven.be/GME/}, chapter 3.

      Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
