\name{Hstarts}
\docType{data}
\alias{Hstarts}
\title{Housing Starts }
\description{
  quarterly observations from 1960-1 to 2001-4

 \emph{number of observations} :  168

 \emph{observation} :  country

   \emph{country} :  Canada

}
\usage{data(Hstarts)}
\format{A time serie containing :
  \describe{
    \item{hs}{ the log of urban housing starts in Canada, not seasonally adjusted, CANSIM series J6001, converted to quarterly}
    \item{hssa}{the log of urban housing starts in Canada, seasonally adjusted, CANSIM series J9001, converted to quarterly. Observations prior to 1966:1 are missing}
  }
}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 13.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
