\name{Icecream}
\docType{data}
\alias{Icecream}
\title{Ice Cream Consumption }
\description{
  four--weekly observations from 1951--03--18 to 1953--07--11

 \emph{number of observations} :  30

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Icecream)}
\format{A time serie containing :
  \describe{
    \item{cons}{consumption of ice cream per head (in pints);}
    \item{income}{average family income per week (in US Dollars);}
    \item{price}{price of ice cream (per pint);}
    \item{temp}{average temperature (in Fahrenheit);}
  }
}
\source{
    Hildreth, C.  and  J.  Lu (1960) \emph{Demand relations with autocorrelated disturbances}, Technical Bulletin No 2765, Michigan State University.

}
\references{
    Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{https://feb.kuleuven.be/GME/}, chapter 4.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
