\name{IncomeUK}
\docType{data}
\alias{IncomeUK}
\title{Seasonally Unadjusted Quarterly Data on Disposable Income and Expenditure }
\description{
  quarterly observations from 1971--1 to 1985--2

 \emph{number of observations} :  58

 \emph{observation} :  country

   \emph{country} :  United Kingdown

}
\usage{data(IncomeUK)}
\format{A time serie containing :
  \describe{
    \item{income}{total disposable income  (million Pounds, current prices)}
    \item{consumption}{consumer expenditure (million Pounds, current prices)}
  }
}
\references{
    Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{http://www.econ.kuleuven.ac.be/GME}, chapters 8 and 9.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}



\keyword{datasets}
