\name{Index.Econometrics}
\docType{documentation}
\alias{Index.Econometrics}
\title{Econometric fields}
\description{
\describe{
\item discrete and limited dependent variables 
 \itemize{
\item \code{\link{Accident}} : Ship Accidents  
\item \code{\link{Doctor}} : Number of Doctor Visits  
\item \code{\link{Fair}} : Extramarital Affairs Data  
\item \code{\link{Hdma}} : The Boston HDMA Data Set  
\item \code{\link{ModeChoice}} : Data to Study Travel Mode Choice  
\item \code{\link{Mroz}} : Labor Supply Data  
\item \code{\link{Participation}} : Labor Force Participation  
\item \code{\link{Strike}} : Strike Duration Data  
\item \code{\link{Tobacco}} : Households Tobacco Budget Share  
}
\item multiple-equations 
 \itemize{
\item \code{\link{Electricity}} : Cost Function for Electricity Producers, 1970  
\item \code{\link{Klein}} : Klein's Model I  
\item \code{\link{ManufCost}} : Manufacturing Costs  
\item \code{\link{Nerlove}} : Cost Function for Electricity Producers, 1955  
}
\item panel 
 \itemize{
\item \code{\link{Airline}} : Cost for U.S. Airlines  
\item \code{\link{Cigarette}} : The Cigarette Consumption Panel Data Set  
\item \code{\link{Fatality}} : Drunk Driving Laws and Traffic Deaths  
\item \code{\link{Grunfeld}} : Grunfeld Investment Data  
\item \code{\link{MunExp}} : Municipal Expenditure Data  
\item \code{\link{SumHes}} : The Penn Table  
}
\item time-series 
 \itemize{
\item \code{\link{CRSPday}} : Daily Returns from the CRSP Database  
\item \code{\link{CRSPmon}} : Monthly Returns from the CRSP Database  
\item \code{\link{DM}} : DM Dollar Exchange Rate  
\item \code{\link{Intratesm}} : Interest Rates (monthly)  
\item \code{\link{Intratesq}} : Interest Rates (quarterly)  
\item \code{\link{LT}} : Dollar Sterling Exchange Rate  
\item \code{\link{Macrodat}} : Macroeconomic Time Series for the United States  
\item \code{\link{Mishkin}} : Inflation and Interest Rates  
\item \code{\link{Mpyr}} : Money, National Product and Interest Rate  
\item \code{\link{Orange}} : The Orange Juice Data Set  
\item \code{\link{Pound}} : Pound-dollar Exchange Rate  
\item \code{\link{Yen}} : Yen-dollar Exchange Rate  
}
}

}
\keyword{documentation}
