\name{Index.Economics}
\docType{documentation}
\alias{Index.Economics}
\title{Economic fields}
\description{
\describe{
\item consumer behaviour 
 \itemize{
\item \code{\link{Cigarette}} : The Cigarette Consumption Panel Data Set  
\item \code{\link{Doctor}} : Number of Doctor Visits  
\item \code{\link{ModeChoice}} : Data to Study Travel Mode Choice  
}
\item environmental economics 
 \itemize{
\item \code{\link{Airq}} : Air Quality for Californian Metropolitan Areas  
\item \code{\link{NaturalPark}} : Willingness to Pay for the Preservation of the Alentejo Natural Park  
}
\item finance 
 \itemize{
\item \code{\link{CRSPday}} : Daily Returns from the CRSP Database  
\item \code{\link{CRSPmon}} : Monthly Returns from the CRSP Database  
\item \code{\link{DM}} : DM Dollar Exchange Rate  
\item \code{\link{Intratesm}} : Interest Rates (monthly)  
\item \code{\link{Intratesq}} : Interest Rates (quarterly)  
\item \code{\link{LT}} : Dollar Sterling Exchange Rate  
\item \code{\link{Pound}} : Pound-dollar Exchange Rate  
\item \code{\link{Yen}} : Yen-dollar Exchange Rate  
}
\item labor economics 
 \itemize{
\item \code{\link{CPSch3}} : Earnings from the Current Population Survey  
\item \code{\link{Earnings}} : Earnings for Three Age Groups  
\item \code{\link{Griliches}} : Wage Datas  
\item \code{\link{Mroz}} : Labor Supply Data  
\item \code{\link{Participation}} : Labor Force Participation  
\item \code{\link{Strike}} : Strike Duration Data  
}
\item macroeconomics 
 \itemize{
\item \code{\link{Consumption}} : Quarterly Data on Consumption and Expenditure  
\item \code{\link{Klein}} : Klein's Model I  
\item \code{\link{Longley}} : The Longley Data  
\item \code{\link{MW}} : Growth of Disposable Income and Treasury Bill Rate  
\item \code{\link{Macrodat}} : Macroeconomic Time Series for the United States  
\item \code{\link{Mishkin}} : Inflation and Interest Rates  
\item \code{\link{Money}} : Money, GDP and Interest Rate in Canada  
\item \code{\link{Mpyr}} : Money, National Product and Interest Rate  
\item \code{\link{Solow}} : Solow's Technological Change Data  
\item \code{\link{SumHes}} : The Penn Table  
\item \code{\link{Tbrate}} : Interest Rate, GDP and Inflation  
}
\item producer behaviour 
 \itemize{
\item \code{\link{Accident}} : Ship Accidents  
\item \code{\link{Electricity}} : Cost Function for Electricity Producers, 1970  
\item \code{\link{Grunfeld}} : Grunfeld Investment Data  
\item \code{\link{Hdma}} : The Boston HDMA Data Set  
\item \code{\link{ManufCost}} : Manufacturing Costs  
\item \code{\link{Metal}} : Production for SIC 33  
\item \code{\link{Nerlove}} : Cost Function for Electricity Producers, 1955  
\item \code{\link{Orange}} : The Orange Juice Data Set  
\item \code{\link{TranspEq}} : Statewide Data on Transportation Equipment Manufacturing  
}
\item public economics 
 \itemize{
\item \code{\link{MunExp}} : Municipal Expenditure Data  
}
\item socio-economics 
 \itemize{
\item \code{\link{Caschool}} : The California Test Score Data Set  
\item \code{\link{Fair}} : Extramarital Affairs Data  
\item \code{\link{Fatality}} : Drunk Driving Laws and Traffic Deaths  
\item \code{\link{MCAS}} : The Massashusets Test Score Data Set  
\item \code{\link{Star}} : Effects on Learning of Small Class Sizes  
}
\item urban and regional economics 
 \itemize{
\item \code{\link{Housing}} : Sales Prices of Houses in the City of Windsor  
}
}

}
\keyword{documentation}
