\name{Index.Observations}
\docType{documentation}
\alias{Index.Observations}
\title{Observations}
\description{
\describe{
\item country 
 \itemize{
\item \code{\link{Consumption}} : Quarterly Data on Consumption and Expenditure  
\item \code{\link{DM}} : DM Dollar Exchange Rate  
\item \code{\link{Garch}} : Daily Observations on Exchange Rates of the Us Dollar Against Other Currencies  
\item \code{\link{Gasoline}} : Gasoline Consumption  
\item \code{\link{Hstarts}} : Housing Starts  
\item \code{\link{Icecream}} : Ice Cream Consumption  
\item \code{\link{IncomeUK}} : Seasonally Unadjusted Quarterly Data on Disposable Income and Expenditure  
\item \code{\link{Intratesm}} : Interest Rates (monthly)  
\item \code{\link{Intratesq}} : Interest Rates (quarterly)  
\item \code{\link{Irates}} : Monthly Interest Rates  
\item \code{\link{Klein}} : Klein's Model I  
\item \code{\link{LT}} : Dollar Sterling Exchange Rate  
\item \code{\link{Longley}} : the Longley Data  
\item \code{\link{MW}} : Growth of Disposable Income and Treasury Bill Rate  
\item \code{\link{Macrodat}} : Macroeconomic Time Series for the United States  
\item \code{\link{ManufCost}} : Manufacturing Costs  
\item \code{\link{Mishkin}} : Inflation and Interest Rates  
\item \code{\link{Money}} : Money, GDP and Interest Rate in Canada  
\item \code{\link{Mpyr}} : Money, National Product and Interest Rate  
\item \code{\link{Orange}} : The Orange Juice Data Set  
\item \code{\link{PE}} : Price and Earnings Index  
\item \code{\link{PPP}} : Exchange Rates and Price Indices for France and Italy  
\item \code{\link{Pound}} : Pound-dollar Exchange Rate  
\item \code{\link{Solow}} : Solow's Technological Change Data  
\item \code{\link{SumHes}} : The Penn Table  
\item \code{\link{Tbrate}} : Interest Rate, GDP and Inflation  
\item \code{\link{Yen}} : Yen-dollar Exchange Rate  
}
\item individuals 
 \itemize{
\item \code{\link{Benefits}} : Unemployement of Blue Collar Workers  
\item \code{\link{Bwages}} : Wages in Belgium  
\item \code{\link{CPSch3}} : Earnings from the Current Population Survey  
\item \code{\link{Doctor}} : Number of Doctor Visits  
\item \code{\link{Earnings}} : Earnings for Three Age Groups  
\item \code{\link{Fair}} : Extramarital Affairs Data  
\item \code{\link{Griliches}} : Wage Datas  
\item \code{\link{Hdma}} : The Boston HDMA Data Set  
\item \code{\link{Males}} : Wages and Education of Young Males  
\item \code{\link{ModeChoice}} : Data to Study Travel Mode Choice  
\item \code{\link{Mroz}} : Labor Supply Data  
\item \code{\link{NaturalPark}} : Willingness to Pay for the Preservation of the Alentejo Natural Park  
\item \code{\link{Participation}} : Labor Force Participation  
\item \code{\link{Schooling}} : Wages and Schooling  
\item \code{\link{Star}} : Effects on Learning of Small Class Sizes  
\item \code{\link{Tobacco}} : Households Tobacco Budget Share  
\item \code{\link{Wages1}} : Wages, Experience and Schooling  
}
\item production units 
 \itemize{
\item \code{\link{Airline}} : Cost for U.S. Airlines  
\item \code{\link{CRSPday}} : Daily Returns from the CRSP Database  
\item \code{\link{CRSPmon}} : Monthly Returns from the CRSP Database  
\item \code{\link{Clothing}} : Sales Data of Men's Fashion Stores  
\item \code{\link{Electricity}} : Cost Function for Electricity Producers, 1970  
\item \code{\link{Grunfeld}} : Grunfeld Investment Data  
\item \code{\link{Labour2}} : Belgian Firms  
\item \code{\link{Nerlove}} : Cost Function for Electricity Producers, 1955  
\item \code{\link{Patents}} : Number of Patents for International Manufacturig Firms  
}
\item regional 
 \itemize{
\item \code{\link{Airq}} : Air Quality for Californian Metropolitan Areas  
\item \code{\link{Cigar}} : Cigarette Consumption  
\item \code{\link{Cigarette}} : The Cigarette Consumption Panel Data Set  
\item \code{\link{Fatality}} : Drunk Driving Laws and Traffic Deaths  
\item \code{\link{Hedonic}} : Hedonic Prices of Cencus Tracts in Boston  
\item \code{\link{Metal}} : Production for SIC 33  
\item \code{\link{MunExp}} : Municipal Expenditure Data  
\item \code{\link{Produc}} : Us States Production  
\item \code{\link{TranspEq}} : Statewide Data on Transportation Equipment Manufacturing  
}
\item schools 
 \itemize{
\item \code{\link{Caschool}} : The California Test Score Data Set  
\item \code{\link{MCAS}} : The Massashusets Test Score Data Set  
}
}

}
\keyword{documentation}
