\name{Index.Source}
\alias{Index.Source}
\title{Source}
\description{
  \describe{
    \itemize{
\item Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}
 \itemize{
\item \code{\link{Bids}} : Bids Received By U.S. Firms
\item \code{\link{BudgetFood}} : Budget Share of Food for Spanish Households
\item \code{\link{BudgetItaly}} : Budget Shares for Italian Households
\item \code{\link{BudgetUK}} : Budget Shares of British Households
\item \code{\link{Car}} : Stated Preferences for Car Choice
\item \code{\link{Computers}} : Prices of Personal Computers
\item \code{\link{Crime}} : Crime in North Carolina
\item \code{\link{Doctor}} : Number of Doctor Visits
\item \code{\link{Earnings}} : Earnings for Three Age Groups
\item \code{\link{HI}} : Health Insurance and Hours Worked By Wives
\item \code{\link{Housing}} : Sales Prices of Houses in the City of Windsor
\item \code{\link{Males}} : Wages and Education of Young Males
\item \code{\link{Mathlevel}} : Level of Calculus Attained for Students Taking Advanced Micro--economics
\item \code{\link{MoneyUS}} : Macroeconomic Series for the United States
\item \code{\link{MunExp}} : Municipal Expenditure Data
\item \code{\link{OFP}} : Visits to Physician Office
\item \code{\link{Oil}} : Oil Investment
\item \code{\link{Participation}} : Labor Force Participation
\item \code{\link{PatentsRD}} : Patents, R\&D and Technological Spillovers for a Panel of Firms
\item \code{\link{Train}} : Stated Preferences for Train Traveling
\item \code{\link{Unemployment}} : Unemployment Duration
\item \code{\link{University}} : Provision of University Teaching and Research
\item \code{\link{Workinghours}} : Wife Working Hours
}

\item Journal of Business Economics and Statistics web site : \url{http://www.amstat.org/publications/jbes/}
 \itemize{
\item \code{\link{Benefits}} : Unemployement of Blue Collar Workers
\item \code{\link{Catsup}} : Choice of Brand for Catsup
\item \code{\link{Cracker}} : Choice of Brand for Crakers
\item \code{\link{Kakadu}} : Willingness to Pay for the Preservation of the Kakadu National Park
\item \code{\link{Ketchup}} : Choice of Brand for Ketchup
\item \code{\link{LaborSupply}} : Wages and Hours Worked
\item \code{\link{Mofa}} : International Expansion of U.S. Mofa's (majority--owned Foreign Affiliates in Fire (finance, Insurance and Real Estate)
\item \code{\link{Somerville}} : Visits to Lake Somerville
\item \code{\link{Tuna}} : Choice of Brand for Tuna
\item \code{\link{Yogurt}} : Choice of Brand for Yogurts
}

\item Journal of Statistics Education's data archive : \url{http://www.amstat.org/publications/jse/jse_data_archive.html}
 \itemize{
\item \code{\link{Diamond}} : Pricing the C's of Diamond Stones
\item \code{\link{FriendFoe}} : Data from the Television Game Show Friend Or Foe ?
}

\item Kenneth Train's home page :  \url{http://elsa.berkeley.edu/~train/}
 \itemize{
\item \code{\link{HC}} : Heating and Cooling System Choice in Newly Built Houses in California
\item \code{\link{Heating}} : Heating System Choice in California Houses
\item \code{\link{Mode}} : Mode Choice
}

\item Baltagi, Badi H. (2003) \emph{Econometric analysis of panel data}, John Wiley and sons, \url{http://www.wiley.com/legacy/wileychi/baltagi/}
 \itemize{
\item \code{\link{Cigar}} : Cigarette Consumption
\item \code{\link{Crime}} : Crime in North Carolina
\item \code{\link{Gasoline}} : Gasoline Consumption
\item \code{\link{Grunfeld}} : Grunfeld Investment Data
\item \code{\link{Hedonic}} : Hedonic Prices of Cencus Tracts in Boston
\item \code{\link{Produc}} : Us States Production
\item \code{\link{Wages}} : Panel Datas of Individual Wages
}

\item Cameron, A.C.  and  P.K.  Trivedi (2005) \emph{Microeconometrics : methods and applications}, Cambridge
 \itemize{
\item \code{\link{DoctorContacts}} : Contacts With Medical Doctor
\item \code{\link{Fishing}} : Choice of Fishing Mode
\item \code{\link{LaborSupply}} : Wages and Hours Worked
\item \code{\link{MedExp}} : Structure of Demand for Medical Care
\item \code{\link{PSID}} : Panel Survey of Income Dynamics
\item \code{\link{PatentsHGH}} : Dynamic Relation Between Patents and R\&D
\item \code{\link{RetSchool}} : Return to Schooling
\item \code{\link{StrikeDur}} : Strikes Duration
\item \code{\link{Treatment}} : Evaluating Treatment Effect of Training on Earnings
\item \code{\link{UnempDur}} : Unemployment Duration
\item \code{\link{VietNamH}} : Medical Expenses in Viet--nam (household Level)
\item \code{\link{VietNamI}} : Medical Expenses in Viet--nam (individual Level)
}

\item Cameron, A.C.  and   Trivedi P.K. (1998) \emph{Regression analysis of count data}, Cambridge University Press, \url{http://cameron.econ.ucdavis.edu/racd/racddata.html}
 \itemize{
\item \code{\link{Bids}} : Bids Received By U.S. Firms
\item \code{\link{DoctorAUS}} : Doctor Visits in Australia
\item \code{\link{OFP}} : Visits to Physician Office
\item \code{\link{PatentsHGH}} : Dynamic Relation Between Patents and R\&D
\item \code{\link{Somerville}} : Visits to Lake Somerville
\item \code{\link{StrikeNb}} : Number of Strikes in Us Manufacturing
}

\item Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, \url{http://www.econ.queensu.ca/ETM/}
 \itemize{
\item \code{\link{CRSPday}} : Daily Returns from the CRSP Database
\item \code{\link{CRSPmon}} : Monthly Returns from the CRSP Database
\item \code{\link{Consumption}} : Quarterly Data on Consumption and Expenditure
\item \code{\link{Doctor}} : Number of Doctor Visits
\item \code{\link{Earnings}} : Earnings for Three Age Groups
\item \code{\link{Hstarts}} : Housing Starts
\item \code{\link{Intratesm}} : Interest Rates (monthly)
\item \code{\link{Intratesq}} : Interest Rates (quarterly)
\item \code{\link{MW}} : Growth of Disposable Income and Treasury Bill Rate
\item \code{\link{Money}} : Money, GDP and Interest Rate in Canada
\item \code{\link{Participation}} : Labor Force Participation
\item \code{\link{Tbrate}} : Interest Rate, GDP and Inflation
}

\item Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}
 \itemize{
\item \code{\link{Accident}} : Ship Accidents
\item \code{\link{Airline}} : Cost for U.S. Airlines
\item \code{\link{Electricity}} : Cost Function for Electricity Producers
\item \code{\link{Fair}} : Extramarital Affairs Data
\item \code{\link{Grunfeld}} : Grunfeld Investment Data
\item \code{\link{Klein}} : Klein's Model I
\item \code{\link{Longley}} : The Longley Data
\item \code{\link{ManufCost}} : Manufacturing Costs
\item \code{\link{Metal}} : Production for SIC 33
\item \code{\link{ModeChoice}} : Data to Study Travel Mode Choice
\item \code{\link{Mroz}} : Labor Supply Data
\item \code{\link{MunExp}} : Municipal Expenditure Data
\item \code{\link{Nerlove}} : Cost Function for Electricity Producers, 1955
\item \code{\link{Solow}} : Solow's Technological Change Data
\item \code{\link{Strike}} : Strike Duration Data
\item \code{\link{TranspEq}} : Statewide Data on Transportation Equipment Manufacturing
}

\item Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, \url{http://www.e.u-tokyo.ac.jp/~hayashi/hayashi_econometrics.htm}
 \itemize{
\item \code{\link{DM}} : DM Dollar Exchange Rate
\item \code{\link{Electricity}} : Cost Function for Electricity Producers
\item \code{\link{Griliches}} : Wage Datas
\item \code{\link{LT}} : Dollar Sterling Exchange Rate
\item \code{\link{Mishkin}} : Inflation and Interest Rates
\item \code{\link{Mpyr}} : Money, National Product and Interest Rate
\item \code{\link{Nerlove}} : Cost Function for Electricity Producers, 1955
\item \code{\link{Pound}} : Pound-dollar Exchange Rate
\item \code{\link{SumHes}} : The Penn Table
\item \code{\link{Yen}} : Yen-dollar Exchange Rate
}

\item Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers, \url{http://wps.aw.com/aw_stockwatsn_economtrcs_1}
 \itemize{
\item \code{\link{CPSch3}} : Earnings from the Current Population Survey
\item \code{\link{Caschool}} : The California Test Score Data Set
\item \code{\link{Cigarette}} : The Cigarette Consumption Panel Data Set
\item \code{\link{Fatality}} : Drunk Driving Laws and Traffic Deaths
\item \code{\link{Hdma}} : The Boston HDMA Data Set
\item \code{\link{Journals}} : Economic Journals Dat Set
\item \code{\link{MCAS}} : The Massashusets Test Score Data Set
\item \code{\link{Macrodat}} : Macroeconomic Time Series for the United States
\item \code{\link{Orange}} : The Orange Juice Data Set
\item \code{\link{Star}} : Effects on Learning of Small Class Sizes
}

\item Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{http://www.econ.kuleuven.ac.be/GME}
 \itemize{
\item \code{\link{Airq}} : Air Quality for Californian Metropolitan Areas
\item \code{\link{Benefits}} : Unemployement of Blue Collar Workers
\item \code{\link{Bwages}} : Wages in Belgium
\item \code{\link{Capm}} : Stock Market Data
\item \code{\link{Clothing}} : Sales Data of Men's Fashion Stores
\item \code{\link{Forward}} : Exchange Rates of US Dollar Against Other Currencies
\item \code{\link{Garch}} : Daily Observations on Exchange Rates of the US Dollar Against Other Currencies
\item \code{\link{Housing}} : Sales Prices of Houses in the City of Windsor
\item \code{\link{Icecream}} : Ice Cream Consumption
\item \code{\link{IncomeUK}} : Seasonally Unadjusted Quarterly Data on Disposable Income and Expenditure
\item \code{\link{Irates}} : Monthly Interest Rates
\item \code{\link{Labour}} : Belgian Firms
\item \code{\link{Males}} : Wages and Education of Young Males
\item \code{\link{MoneyUS}} : Macroeconomic Series for the United States
\item \code{\link{NaturalPark}} : Willingness to Pay for the Preservation of the Alentejo Natural Park
\item \code{\link{PE}} : Price and Earnings Index
\item \code{\link{PPP}} : Exchange Rates and Price Indices for France and Italy
\item \code{\link{PatentsRD}} : Patents, R\&D and Technological Spillovers for a Panel of Firms
\item \code{\link{Pricing}} : Returns of Size-based Portfolios
\item \code{\link{SP500}} : Returns on Standard \& Poor's 500 Index
\item \code{\link{Schooling}} : Wages and Schooling
\item \code{\link{Tobacco}} : Households Tobacco Budget Share
\item \code{\link{Wages1}} : Wages, Experience and Schooling
}

}
}

}

\keyword{documentation}
