\name{Intratesm}
\docType{data}
\alias{Intratesm}
\title{Interest Rates (monthly) }
\description{
  monthly observations from 1955-1 to 2001-12

 \emph{number of observations} :  564

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Intratesm)}
\format{A time serie containing :
  \describe{
    \item{rff}{the federal funds rate}
    \item{rt1}{the 1 year treasury bill rate}
    \item{rt3}{the 3 year treasury constant maturity rate}
    \item{rt10}{the 10 year treasury constant maturity rate}
  }
}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, \url{http://www.econ.queensu.ca/ETM/}, chapter 13 and 14.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
