\name{Irates}
\docType{data}
\alias{Irates}
\title{Monthly Interest Rates }
\description{
  monthly observations from 1946--12 to 1991--02

 \emph{number of observations} :  531

 \emph{observation} :  country

   \emph{country} :  United--States

}
\usage{data(Irates)}
\format{A time series containing :
  \describe{
    \item{r1}{interest rate for a maturity of 1 months  (\% per year). }
    \item{r2}{interest rate for a maturity of 2 months  (\% per year). }
    \item{r3}{interest rate for a maturity of 3 months  (\% per year). }
    \item{r5}{interest rate for a maturity of 5 months  (\% per year). }
    \item{r6}{interest rate for a maturity of 6 months  (\% per year). }
    \item{r11}{interest rate for a maturity of 11 months  (\% per year). }
    \item{r12}{interest rate for a maturity of 12 months  (\% per year). }
    \item{r36}{interest rate for a maturity of 36 months  (\% per year). }
    \item{r60}{interest rate for a maturity of 60 months  (\% per year). }
    \item{r120}{interest rate for a maturity of 120 months  (\% per year). }
  }
}
\source{
    McCulloch, J.H.  and  H.C.  Kwon (1993) \emph{U.S. term structure data, 1947--1991}, Ohio State Working Paper 93-6, Ohio State University, Columbus.

}
\references{
    Verbeek, Marno (2004) \emph{A Guide to Modern Econometrics}, John Wiley and Sons, chapter 8.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
