\name{Journals}
\docType{data}
\alias{Journals}
\title{Economic Journals Data Set }
\description{
  a cross-section from 2000

 \emph{number of observations} :  180

 \emph{observation} :  goods

}
\usage{data(Journals)}
\format{A dataframe containing :
  \describe{
    \item{title}{journal title}
    \item{pub}{publisher}
    \item{society}{ scholarly society ?}
    \item{libprice}{library subscription price}
    \item{pages}{number of pages}
    \item{charpp}{characters per page}
    \item{citestot}{total number of citations}
    \item{date1}{year journal was founded}
    \item{oclc}{number of library subscriptions}
    \item{field}{field description}
  }
}
\source{
    Professor Theodore Bergstrom of the Department of Economics at the University of California, San Diego.

}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers, chapter 6.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
