\name{Klein}
\docType{data}
\alias{Klein}
\title{Klein's Model I }
\description{
  annual observations from 1920 to 1941

 \emph{number of observations} :  22

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Klein)}
\format{A time series containing :
  \describe{
    \item{cons}{consumption}
    \item{profit}{corporate profits}
    \item{privwage}{private wage bill}
    \item{inv}{investment}
    \item{lcap}{previous year's capital stock}
    \item{gnp}{GNP}
    \item{pubwage}{government wage bill}
    \item{govspend}{government spending}
    \item{taxe}{taxes}
  }
}
\source{
    Klein, L. (1950) \emph{Economic fluctuations in the United States, 1921-1941}, New York, John Wiley and Sons.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}, Table F15.1.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
