\name{Labour2}
\docType{data}
\alias{Labour2}
\title{Belgian Firms }
\description{
  a cross-section from 1996

 \emph{number of observations} :  569

 \emph{observation} :  production units

   \emph{country} :  Belgium

}
\usage{data(Labour2)}
\format{A dataframe containing :
  \describe{
    \item{capital}{total fixed assets, end of 1995 (in 1000000 euro)}
    \item{labour}{number of workers (employment) }
    \item{output}{value added (in 1000000 euro)}
    \item{wage}{wage costs per worker (in 1000 euro)}
  }
}
\references{
    Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{http://www.econ.kuleuven.ac.be/GME}, chapter 4.

}
\seealso{\code{\link{Index.Books}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}

\keyword{datasets}
