\name{Longley}
\docType{data}
\alias{Longley}
\title{The Longley Data }
\description{
  annual observations from 1947 to 1962

 \emph{number of observations} :  16

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Longley)}
\format{A time series containing :
  \describe{
    \item{employ}{employment (1,000s)}
    \item{price}{GNP deflator}
    \item{gnp}{nominal GNP (millions)}
    \item{armed}{armed forces}
  }
}
\source{
    Longley, J. (1967) \dQuote{An appraisal of least squares programs from the point of view of the user}, \emph{Journal of the American Statistical Association}, \bold{62}, 819-841.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{https://archive.org/details/econometricanaly0000gree_f4x3}, Table F4.2.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
