\name{LT}
\docType{data}
\alias{LT}
\title{Dollar Sterling Exchange Rate }
\description{
  annual observations from 1791 to 1990

 \emph{number of observations} :  200

 \emph{observation} :  country

   \emph{country} :  United Kingdown

}
\usage{data(LT)}
\format{A time serie containing :
  \describe{
    \item{s}{US *Dollar / *Pound exchange rate}
    \item{uswpi}{US wholesale price index, normalized to 100 for 1914}
    \item{ukwpi}{US wholesale price index, normalized to 100 for 1914}
  }
}
\source{
    Lothian, J.  and  M.  Taylor (1996) \dQuote{Real exchange rate behavior: the recent float from the perspective of the past two centuries}, \emph{Journal of Political Economy}, \bold{104}, 488-509.

}
\references{
    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, \url{http://fhayashi.fc2web.com/hayashi_econometrics.htm}, chapter 9, 613-621.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
