\name{Macrodat}
\docType{data}
\alias{Macrodat}
\title{Macroeconomic Time Series for the United States }
\description{
  quarterly observations from 1959-1 to 2000-4

 \emph{number of observations} :  168

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Macrodat)}
\format{A time serie containing :
  \describe{
    \item{lhur}{unemployment rate (average of months in quarter)       }
    \item{punew}{cPI   (Average of Months in Quarter)       }
    \item{fyff}{federal funds interest rate (last month in quarter)       }
    \item{fygm3}{3 month treasury bill interest rate (last month in quarter)       }
    \item{fygt1}{1 year treasury bond interest rate (last month in quarter)       }
    \item{exruk}{dollar / Pound exchange rate (last month in quarter)       }
    \item{gdpjp}{real GDP for Japan       }
  }
}
\source{
    Bureau of Labor Statistics, OECD, Federal Reserve.

}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers, \url{http://wps.aw.com/aw_stockwatsn_economtrcs_1}, chapter 12 and 14.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
