\name{ManufCost}
\docType{data}
\alias{ManufCost}
\title{Manufacturing Costs }
\description{
  annual observations from 1947 to 1971

 \emph{number of observations} :  25

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(ManufCost)}
\format{A time series containing :
  \describe{
    \item{cost}{cost index}
    \item{sk}{capital cost share}
    \item{sl}{labor cost share}
    \item{se}{energy cost share}
    \item{sm}{materials cost share}
    \item{pk}{capital price}
    \item{pl}{labor price}
    \item{pe}{energy price}
    \item{pm}{materials price}
  }
}
\source{
    Berndt, E.  and  D.  Wood (1975) \dQuote{Technology, prices and the derived demand for energy}, \emph{Journal of Economics and Statistics}, \bold{57}, 376-384.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}, Table F14.1.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
