\name{Metal}
\docType{data}
\alias{Metal}
\title{Production for SIC 33 }
\description{
  a cross-section

 \emph{number of observations} :  27

 \emph{observation} :  regional

   \emph{country} :  United States

}
\usage{data(Metal)}
\format{A dataframe containing :
  \describe{
    \item{va}{output}
    \item{labor}{labor input}
    \item{capital}{capital input}
  }
}
\source{
    Aigner, D.,  K.   Lovell and  P.  Schmidt (1977) \dQuote{Formulation and estimation of stochastic frontier production models}, \emph{Journal of Econometrics}, \bold{6}, 21-37.

    Hildebrand, G.  and  T.  Liu (1957) \emph{Manufacturing production functions in the United States}, Ithaca, N.Y.: Cornell University Press.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{https://archive.org/details/econometricanaly0000gree_f4x3}, Table F6.1.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
