\name{Mishkin}
\docType{data}
\alias{Mishkin}
\title{Inflation and Interest Rates }
\description{
  monthly observations from 1950-2 to 1990-12

 \emph{number of observations} :  491

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Mishkin)}
\format{A time serie containing :
  \describe{
    \item{pai1}{one-mounth inflation rate (in percent, annual rate)}
    \item{pai3}{three-mounth inflation rate (in percent, annual rate)}
    \item{tb1}{one-mounth T-bill rate (in percent, annual rate)}
    \item{tb3}{three-mounth T-bill rate (in percent, annual rate)}
    \item{cpi}{CPI for urban consumers, all items (the 1982-1984 average is set to 100)}
  }
}
\source{
    Mishkin, F. (1992) \dQuote{Is the Fisher effect for real ?}, \emph{Journal of Monetary Economics}, \bold{30}, 195-215.

}
\references{
    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, \url{http://fhayashi.fc2web.com/hayashi_econometrics.htm}, chapter 2, 176-184.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
