\name{ModeChoice}
\docType{data}
\alias{ModeChoice}
\title{Data to Study Travel Mode Choice }
\description{
  a cross-section

 \emph{number of observations} :  840

 \emph{observation} :  individuals

   \emph{country} :  Australia

}
\usage{data(ModeChoice)}
\format{A dataframe containing :
  \describe{
    \item{mode}{choice : air, train, bus or car}
    \item{ttme}{terminal waiting cost time, 0 for car}
    \item{invc}{in vehicle cost-cost component}
    \item{invt}{travel time in vehicle}
    \item{gc}{generalized cost measure}
    \item{hinc}{household income}
    \item{psize}{party size in mode chosen}
  }
}
\source{
    Greene, W.H.  and  D.  Hensher (1997) \emph{Multinomial logit and discrete choice models} \emph{in}  Greene, W. H. (1997) \emph{LIMDEP version 7.0 user's manual revised}, Plainview, New York econometric software, Inc .

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}, Table F21.2.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
