\name{Money}
\docType{data}
\alias{Money}
\title{Money, GDP and Interest Rate in Canada }
\description{
  quarterly observations from 1967-1 to 1998-4

 \emph{number of observations} :  128

 \emph{observation} :  country

   \emph{country} :  Canada

}
\usage{data(Money)}
\format{A time serie containing :
  \describe{
    \item{m}{log of the real money supply}
    \item{y}{the log of GDP, in 1992 dollars, seasonally adjusted}
    \item{p}{the log of the price level}
    \item{r}{the 3-month treasury till rate}
  }
}
\source{
    CANSIM Database of Statistics Canada.

}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, \url{http://www.econ.queensu.ca/ETM/}, chapter 7 and 8.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
