\name{MoneyUS}
\docType{data}
\alias{MoneyUS}
\title{Macroeconomic Series for the United States }
\description{
  quarterly observations from 1954--01 to 1994--12

 \emph{number of observations} :  164

   \emph{country} :  United States

}
\usage{data(MoneyUS)}
\format{A time serie containing :
  \describe{
    \item{m}{log of real M1 money stock}
    \item{infl}{quaterly inflation rate (change in log prices), \% per year}
    \item{cpr}{commercial paper rate, \% per year}
    \item{y}{log real GDP (in billions of 1987 dollars)}
    \item{tbr}{treasury bill rate}
  }
}
\source{
    Hoffman, D.L.  and  R.H.  Rasche (1996) \dQuote{Assessing forecast performance in a cointegrated system}, \emph{Journal of Applied Econometrics}, \bold{11}, 495--517.

}
\references{
    Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{http://www.econ.kuleuven.ac.be/GME}, chapter 9.

      Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
