\name{MunExp}
\docType{data}
\alias{MunExp}
\title{Municipal Expenditure Data }
\description{
  a panel of 265 observations  from 1979  to 1987

 \emph{number of observations} :  2385

 \emph{observation} :  regional

   \emph{country} :  Sweden

}
\usage{data(MunExp)}
\format{A dataframe containing :
  \describe{
    \item{id}{identification}
    \item{year}{date}
    \item{expend}{expenditure}
    \item{revenue}{revenue from taxes and fees}
    \item{grants}{grants from Central Government}
  }
}
\source{
    Dahlberg, M. and E. Johansson (2000) \dQuote{An examination of the dynamic behavior of local government using GMM boot-strapping methods}, \emph{Journal of Applied Econometrics}, \bold{21}, 333-355.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}, Table F18.1.

      Journal of Applied Econometrics data archive : \url{http://qed.econ.queensu.ca/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
