\name{MW}
\docType{data}
\alias{MW}
\title{Growth of Disposable Income and Treasury Bill Rate }
\description{
  quarterly observations from 1963-3 to 1975-4

 \emph{number of observations} :  50

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(MW)}
\format{A time serie containing :
  \describe{
    \item{rdi}{the rate of growth of real U.S. disposable income, seasonally adjusted}
    \item{trate}{the U.S. treasury bill rate}
  }
}
\source{
    MacKinnon, J. G.  and  H. T.  White (1985) \dQuote{Some heteroskedasticity consistent  covariance matrix estimators with improved finite sample properties}, \emph{Journal of Econometrics}, \bold{29}, 305-325.

}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, \url{http://www.econ.queensu.ca/ETM/}, chapter 5.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
