\name{Nerlove}
\docType{data}
\alias{Nerlove}
\title{Cost Function for Electricity Producers, 1955 }
\description{
  a cross-section from 1955 to 1955

 \emph{number of observations} :  159

 \emph{observation} :  production units

   \emph{country} :  United States

}
\usage{data(Nerlove)}
\format{A dataframe containing :
  \describe{
    \item{cost}{total cost}
    \item{output}{total output}
    \item{pl}{wage rate}
    \item{sl}{cost share for labor}
    \item{pk}{capital price index}
    \item{sk}{cost share for capital}
    \item{pf}{fuel price}
    \item{sf}{cost share for fuel}
  }
}
\source{
    Nerlove, M. (1963) \emph{Returns to scale in electricity industry} \emph{in}  Christ, C. ed. (1963) \emph{Measurment in economics: studies in mathematical economics and econometrics in memory of Yehuda Grunfeld }, Standford, California, Standford University Press .

    Christensen, L.  and  W. H.  Greene (1976) \dQuote{Economies of scale in U.S. electric power generation}, \emph{Journal of Political Economy}, \bold{84}, 655-676.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}, Table F14.2.

    Hayashi, F. (2000) \emph{Econometrics}, Princeton University Press, \url{http://fhayashi.fc2web.com/hayashi_econometrics.htm}, chapter 1, 76-84.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
