\name{Orange}
\docType{data}
\alias{Orange}
\title{The Orange Juice Data Set }
\description{
  monthly observations from 1948-01 to 2001-06

 \emph{number of observations} :  642

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Orange)}
\format{A time serie containing :
  \describe{
    \item{priceoj}{producer price for frozen orange juice}
    \item{pricefg}{producer price index for finished goods}
    \item{fdd}{freezing degree days (from daily minimimum temperature reacorded at Orlando area airports)}
  }
}
\source{
    U.S. Bureau of Labor Statistics for PPIOJ and PWFSA, National Oceanic and Atmospheric Administration (NOAA) of the U.S Department of Commerce for FDD.

}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers, \url{http://wps.aw.com/aw_stockwatsn_economtrcs_1}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
