\name{Participation}
\docType{data}
\alias{Participation}
\title{Labor Force Participation }
\description{
  a cross-section

 \emph{number of observations} :  872

 \emph{observation} :  individuals

   \emph{country} :  Switzerland

}
\usage{data(Participation)}
\format{A dataframe containing :
  \describe{
    \item{lfp}{labour force participation ?}
    \item{lnnlinc}{the log of nonlabour income}
    \item{age}{age in years divided by 10}
    \item{educ}{years of formal education}
    \item{nyc}{the number of young children (younger than 7)}
    \item{noc}{ number of older children}
    \item{foreign}{  foreigner ?}
  }
}
\source{
     Gerfin, Michael (1996) \dQuote{Parametric and semiparametric estimation of the binary  response}, \emph{Journal of Applied Econometrics}, \bold{11(3)}, 321-340.

}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 11.

      Journal of Applied Econometrics data archive : \url{http://qed.econ.queensu.ca/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
