\name{Patents}
\docType{data}
\alias{Patents}
\title{Number of Patents for International Manufacturig Firms }
\description{
  a cross-section from 1990--1991

 \emph{number of observations} :  181

 \emph{observation} :  production units

   \emph{country} :  world

}
\usage{data(Patents)}
\format{A dataframe containing :
  \describe{
    \item{p91}{number of patents in 1991}
    \item{p90}{number of patents in 1990}
    \item{lr91}{log R&D expenditures 1991}
    \item{lr90}{log R&D expenditures 1990}
    \item{industry}{a factor with levels (aerospace,chemistry,computer,machinery\_and\_instruments,motor\_vehicles)}
    \item{country}{a factor with levels (japan,us,other)}
  }
}
\source{
    Cincera, M. (1997) \dQuote{Patents, R\&D, and technological spillovers at the firm level : some evidence from econometric count models for panel data}, \emph{Journal of Applied Econometrics }, \bold{12}, 265--280.

}
\references{
    Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{http://www.econ.kuleuven.ac.be/GME}, chapter 7.

}
\seealso{\code{\link{Index.Books}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}

\keyword{datasets}
