\name{PPP}
\docType{data}
\alias{PPP}
\title{Exchange Rates and Price Indices for France and Italy }
\description{
  monthly observations from 1981--01 to 1996--06

 \emph{number of observations} :  186

 \emph{observation} :  country

   \emph{country} :  France and Italy

}
\usage{data(PPP)}
\format{A time serie containing :
  \describe{
    \item{lnit}{log price index italy}
    \item{lnfr}{log price index france}
    \item{lnx}{ log exchange rate france/italy}
    \item{cpiit}{consumer price index italy}
    \item{cpifr}{consumer price index france}
  }
}
\source{
    Datastream.

}
\references{
    Verbeek, Marno (2004) \emph{A guide to modern econometrics}, John Wiley and Sons, \url{http://www.econ.kuleuven.ac.be/GME}, chapters 8 and 9.

}
\seealso{\code{\link{Index.Books}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}, \code{\link{Index.Time.Series}}}

\keyword{datasets}
