\name{Pricing}
\docType{data}
\alias{Pricing}
\title{Returns of Size-based Portfolios }
\description{
  monthly observations from 1959--02 to 1993--11

 \emph{number of observations} :  418

}
\usage{data(Pricing)}
\format{A time serie containing :
  \describe{
    \item{r1}{monthly return on portfolio 1 (small firms)}
    \item{r2}{monthly return on portfolio 2 }
    \item{r3}{monthly return on portfolio 3}
    \item{r4}{monthly return on portfolio 4 }
    \item{r5}{monthly return on portfolio 5 }
    \item{r6}{monthly return on portfolio 6 }
    \item{r7}{monthly return on portfolio 7 }
    \item{r8}{monthly return on portfolio 8 }
    \item{r9}{monthly return on portfolio 9 }
    \item{r10}{monthly return on portfolio 10 (large firms)}
    \item{rf}{risk free rate (return on 3-month T-bill)}
    \item{cons}{real per capita consumption growth based on total US personal consumption expenditures (nondurables and services)}
  }
}
\source{
    Center for research in security prices.

}
\references{
    Verbeek, Marno (2004) \emph{A Guide to Modern Econometrics}, John Wiley and Sons, chapter 5.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
