\name{Produc}
\docType{data}
\alias{Produc}
\title{Us States Production }
\description{
  a panel of 48 observations  from 1970   to 1986

 \emph{number of observations} :  816

 \emph{observation} :  regional

   \emph{country} :  United States

}
\usage{data(Produc)}
\format{A dataframe containing :
  \describe{
    \item{state}{the state}
    \item{year}{the year}
    \item{pcap}{private capital stock}
    \item{hwy}{highway and streets}
    \item{water}{water and sewer facilities}
    \item{util}{other public buildings and structures}
    \item{pc}{public capital}
    \item{gsp}{gross state products}
    \item{emp}{
      labor input measured by the employment in 
      non--agricultural payrolls
    }
    \item{unemp}{state unemployment rate}
  }
}
\source{
    Munnell, A. (1990) \dQuote{Why has productivity growth declined? Productivity and public investment}, \emph{New England Economic Review}, 3--22.

    Baltagi, B. H.  and  N.  Pinnoi (1995) \dQuote{Public capital stock and state productivity growth: further evidence}, \emph{Empirical Economics}, \bold{20}, 351--359.

}
\references{
    Baltagi, Badi H. (2003) \emph{Econometric analysis of panel data}, John Wiley and sons, \url{https://www.wiley.com/legacy/wileychi/baltagi/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
