\name{PSID}
\docType{data}
\alias{PSID}
\title{Panel Survey of Income Dynamics }
\description{
  a cross-section from 1993

 \emph{number of observations} :  4856

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(PSID)}
\format{A dataframe containing :
  \describe{
    \item{intnum}{1968 interview number}
    \item{persnum}{person number}
    \item{age}{age of individual}
    \item{educatn}{highest grade completed}
    \item{earnings}{total labor income}
    \item{hours}{annual work hours}
    \item{kids}{live births to this individual}
    \item{married}{last known marital status (married, never married, windowed, divorced, separated, NA/DF, no histories)}
  }
}
\source{
    Panel Survey of Income Dynamics.

}
\references{
    Cameron, A.C.  and  P.K.  Trivedi (2005) \emph{Microeconometrics : methods and applications}, Cambridge, pp. 295--300.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
