\name{RetSchool}
\docType{data}
\alias{RetSchool}
\title{Return to Schooling }
\description{
  a panel of 48 observations  from 1970   to 1986

 \emph{number of observations} :  5225

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(RetSchool)}
\format{A time serie containing :
  \describe{
    \item{wage76}{wage in 1876}
    \item{grade76}{grade level in 1976}
    \item{exp76}{experience 1n 1976}
    \item{black}{black ?}
    \item{south76}{lived in south in 1976 ?}
    \item{smsa76}{lived in smsa in 1976 ?}
    \item{region}{region, a factor with levels (un,midatl,enc,wnc,sa,esc,wsc,m,p)}
    \item{smsa66}{lived in smsa in 1966 ?}
    \item{momdad14}{lived with both parents at age 14 ?}
    \item{sinmom14}{lived with mother only at age 14 ?}
    \item{nodaded}{father has no formal eduction ?}
    \item{nomomed}{mother has no formal education ?}
    \item{daded}{mean grade level of father}
    \item{momed}{mean grade level of mother}
    \item{famed}{father's and mother's education, a factor with 9 levels}
    \item{age76}{age in 1976}
    \item{col4}{is any 4-year college nearby ?}
  }
}
\source{
    Kling, Jeffrey R. (2001) \dQuote{Interpreting Instrumental Variables Estimates of the Return to Schooling}, \emph{Journal of Business and Economic Statistics}, \bold{19(3)}, july, 358--364.

    Dehejia, R.H.  and  S.  Wahba (2002) \dQuote{Propensity-score Matching Methods for Nonexperimental Causal Studies}, \emph{Restat}, 151--161.

}
\references{
    Cameron, A.C.  and  P.K.  Trivedi (2005) \emph{Microeconometrics : methods and applications}, Cambridge.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
