\name{Solow}
\docType{data}
\alias{Solow}
\title{Solow's Technological Change Data }
\description{
  annual observations from 1909 to 1949

 \emph{number of observations} :  41

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Solow)}
\format{A time serie containing :
  \describe{
    \item{q}{output}
    \item{k}{capital/labor ratio}
    \item{A}{index of technology}
  }
}
\source{
    Solow, R. (1957) \dQuote{Technical change and the aggregate production function}, \emph{Review of Economics and Statistics}, \bold{39}, 312-320.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}, Table F7.2.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
