\name{Somerville}
\docType{data}
\alias{Somerville}
\title{Visits to Lake Somerville }
\description{
  a cross-section from 1980

 \emph{number of observations} :  659

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Somerville)}
\format{A dataframe containing :
  \describe{
    \item{visits}{
      annual number of visits to lake Somerville
    }
    \item{quality}{
      quality ranking score for lake Somerville
    }
    \item{ski}{
      engaged in water--skiing at the lake ?
    }
    \item{income}{annual household income}
    \item{feeSom}{
      annual user fee paid at lake Somerville ?
    }
    \item{costCon}{
      expenditures when visiting lake Conroe
    }
    \item{costSom}{
      expenditures when visiting lake Somerville
    }
    \item{costHoust}{
      expenditures when visiting lake Houston
    }
  }
}
\source{
    Seller, Christine, John R. Stoll and Jean--Paul Chavas (1985) \dQuote{Valuation of empirical measures of welfare change : a comparison of nonmarket techniques}, \emph{Land Economics}, \bold{61(2)}, May, 156--175.

    Gurmu, Shiferaw  and  Pravin K.  Trivedi (1996) \dQuote{ Excess zeros in count models for recreational trips}, \emph{Journal of Business and Economics Statistics}, \bold{14(4)}, October, 469--477.

    Santos Silva, Jao M. C. (2001) \dQuote{A score test for non--nested hypotheses with applications to discrete data models}, \emph{Journal of Applied Econometrics}, \bold{16(5)}, 577--597.

}
\references{
      Journal of Business Economics and Statistics web site : \url{https://amstat.tandfonline.com/loi/ubes20}.
    Cameron, A.C.  and   Trivedi P.K. (1998) \emph{Regression analysis of count data}, Cambridge University Press, \url{https://cameron.econ.ucdavis.edu/racd/racddata.html}, chapter 6.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}



\keyword{datasets}
