\name{Strike}
\docType{data}
\alias{Strike}
\title{Strike Duration Data }
\description{
  a cross-section from 1968 to 1976

 \emph{number of observations} :  62

   \emph{country} :  United States

}
\usage{data(Strike)}
\format{A dataframe containing :
  \describe{
    \item{duration}{strike duration in days}
    \item{prod}{unanticipated output}
  }
}
\source{
    Kennan, J. (1985) \dQuote{The duration of contract strikes in U.S. manufacturing}, \emph{Journal of Econometrics}, \bold{28}, 5-28.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{http://www.prenhall.com/greene/greene1.html}, Table F22.1.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
