\name{StrikeDur}
\docType{data}
\alias{StrikeDur}
\title{Strikes Duration }
\description{
  a cross-section from 1968 to 1976

 \emph{number of observations} :  566

   \emph{country} :  United States

}
\usage{data(StrikeDur)}
\format{A dataframe containing :
  \describe{
    \item{dur}{duration of the strike in days}
    \item{gdp}{measure of stage of business cycle (deviation of monthly log industrial production in manufacturing from prediction from OLS on time, time-squared and monthly dummies)}
  }
}
\source{
    Kennan, J. (1985) \dQuote{The Duration of Contract strikes in U.S. Manufacturing}, \emph{Journal of Econometrics}, \bold{28}, 5-28.

}
\references{
    Cameron, A.C.  and  P.K.  Trivedi (2005) \emph{Microeconometrics : methods and applications}, Cambridge, pp. 574--5 and 582.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
