\name{StrikeNb}
\docType{data}
\alias{StrikeNb}
\title{Number of Strikes in Us Manufacturing }
\description{
  monthly observations from 1968(1) to 1976 (12)

 \emph{number of observations} :  108

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(StrikeNb)}
\format{A time serie containing :
  \describe{
    \item{strikes}{number of strikes (number of contract strikes in U.S. manufacturing beginning each month)}
    \item{output}{ level of economic activity (measured as cyclical departure of aggregate production from its trend level)}
    \item{time}{ a time trend from 1 to 108}
  }
}
\source{
    Kennan, J. (1985) \dQuote{The Duration of Contract strikes in U.S. Manufacturing}, \emph{Journal of Econometrics}, \bold{28}, 5-28.

    Cameron, A.C.  and    Trivedi P.K. (1990) \dQuote{Regression Based Tests for Overdispersion in the Poisson  Model}, \emph{Journal of Econometrics}, december, 347-364.

}
\references{
    Cameron, A.C.  and   Trivedi P.K. (1998) \emph{Regression analysis of count data}, Cambridge University Press, \url{http://cameron.econ.ucdavis.edu/racd/racddata.html}, chapter 7.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
