\name{Tbrate}
\docType{data}
\alias{Tbrate}
\title{Interest Rate, GDP and Inflation }
\description{
  quarterly observations from 1950-1 to 1996-4

 \emph{number of observations} :  188

 \emph{observation} :  country

   \emph{country} :  Canada

}
\usage{data(Tbrate)}
\format{A time serie containing :
  \describe{
    \item{r}{the 91-day treasury bill rate}
    \item{y}{the log of real GDP}
    \item{pi}{ the inflation rate}
  }
}
\source{
    CANSIM database of Statistics Canada.

}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 2.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
