\name{TranspEq}
\docType{data}
\alias{TranspEq}
\title{
  Statewide Data on Transportation Equipment Manufacturing 
}
\description{
  a cross-section

 \emph{number of observations} :  25

 \emph{observation} :  regional

   \emph{country} :  United States

}
\usage{data(TranspEq)}
\format{A dataframe containing :
  \describe{
    \item{state}{state name}
    \item{va}{output}
    \item{capital}{capital input}
    \item{labor}{labor input}
    \item{nfirm}{number of firms}
  }
}
\source{
    Zellner, A.  and  N.  Revankar (1970) \dQuote{Generalized production functions}, \emph{Review of Economic Studies}, \bold{37}, 241-250.

}
\references{
    Greene, W.H. (2003) \emph{Econometric Analysis}, Prentice Hall, \url{https://archive.org/details/econometrics0000haya}, Table F9.2.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
