\name{Tuna}
\docType{data}
\alias{Tuna}
\title{Choice of Brand for Tuna }
\description{
  a cross-section

 \emph{number of observations} :  13705

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Tuna)}
\format{A dataframe containing :
  \describe{
    \item{hid}{individuals identifiers}
    \item{id}{purchase identifiers}
    \item{choice}{
      one of \code{skw} (Starkist water), 
      \code{cosw} (Chicken of the sea water), 
      \code{pw} (store--specific private 
      label water), \code{sko} (Starkist 
      oil), \code{coso} (Chicken of the sea 
      oil)}
    \item{price.z}{price of brand z}
  }
}
\source{
    Kim, Byong--Do,  Robert C.   Blattberg and  Peter E.  Rossi (1995) \dQuote{Modeling the distribution of price sensitivity and implications for optimal retail pricing}, \emph{Journal of Business Economics and Statistics}, \bold{13(3)}, 291.

}
\references{
      Journal of Business Economics and Statistics web site : \url{https://amstat.tandfonline.com/loi/ubes20}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
