\name{University}
\docType{data}
\alias{University}
\title{Provision of University Teaching and Research }
\description{
  a cross-section from 1988

 \emph{number of observations} :  62

 \emph{observation} :  schools

   \emph{country} :  United Kingdown

}
\usage{data(University)}
\format{A dataframe containing :
  \describe{
    \item{undstudents}{undergraduate students}
    \item{poststudents}{postgraduate students}
    \item{nassets}{net assets}
    \item{acnumbers}{academic numbers}
    \item{acrelnum}{academic related numbers}
    \item{clernum}{clerical numbers}
    \item{compop}{computer operators}
    \item{techn}{technicians}
    \item{stfees}{student fees}
    \item{acpay}{academic pay}
    \item{acrelpay}{academic related pay}
    \item{secrpay}{secretarial pay}
    \item{admpay}{admin pay}
    \item{agresrk}{aggregate research rank}
    \item{furneq}{furniture and equipment}
    \item{landbuild}{land and buildings}
    \item{resgr}{research grants}
  }
}
\source{
    Glass, J.C.,  D.G.   McKillop and  N.  Hyndman (1995) \dQuote{Efficiency in the provision of university teaching and research : an empirical analysis of UK universities}, \emph{Journal of Applied Econometrics}, \bold{10(1)}, january--march, 61--72.

}
\references{
      Journal of Applied Econometrics data archive : \url{http://qed.econ.queensu.ca/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
