\name{VietNamH}
\docType{data}
\alias{VietNamH}
\title{Medical Expenses in Viet--nam (household Level) }
\description{
  a cross-section from 1997

 \emph{number of observations} :  5999

 \emph{observation} :  households

   \emph{country} :  Vietnam

}
\usage{data(VietNamH)}
\format{A dataframe containing :
  \describe{
    \item{sex}{gender of household head (male,female)}
    \item{age}{age of household head}
    \item{educyr}{ schooling year of household head}
    \item{farm}{farm household ?}
    \item{urban}{urban household ?}
    \item{hhsize}{  household size}
    \item{lntotal}{log household total expenditure}
    \item{lnmed}{ log household medical expenditure}
    \item{lnrlfood}{log household food expenditure}
    \item{lnexp12m}{log of total household health care expenditure for 12 months}
    \item{commune}{commune}
  }
}
\source{
    Vietnam World Bank Livings Standards Survey.

}
\references{
    Cameron, A.C.  and  P.K.  Trivedi (2005) \emph{Microeconometrics : methods and applications}, Cambridge, pp.88--90.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
