\name{VietNamI}
\docType{data}
\alias{VietNamI}
\title{Medical Expenses in Vietnam (individual Level) }
\description{
  a cross-section from 1997

 \emph{number of observations} :  27765

 \emph{observation} :  individuals

   \emph{country} :  Vietnam

}
\usage{data(VietNamI)}
\format{A dataframe containing :
  \describe{
    \item{pharvis}{number of direct pharmacy visits}
    \item{lnhhexp}{log of total medical expenditure }
    \item{age}{age of household head}
    \item{sex}{gender (male,female)}
    \item{married}{married ?}
    \item{educ}{completed diploma level ?}
    \item{illness}{number of of illnesses experiences in past 12 months}
    \item{injury}{injured during survey period ?}
    \item{illdays}{number of illness days}
    \item{actdays}{number of days of limited activity}
    \item{insurance}{respondent has health insurance coverage ?}
    \item{commune}{commune}
  }
}
\source{
    Vietnam World Bank Livings Standards Survey.

}
\references{
    Cameron, A.C.  and  P.K.  Trivedi (2005) \emph{Microeconometrics : methods and applications}, Cambridge, pp.848--853.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
