\name{Workinghours}
\docType{data}
\alias{Workinghours}
\title{Wife Working Hours }
\description{
  a cross-section from 1987

 \emph{number of observations} :  3382

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Workinghours)}
\format{A dataframe containing :
  \describe{
    \item{hours}{wife working hours per year}
    \item{income}{the other household income in hundreds of dollars}
    \item{age}{age of the wife}
    \item{education}{education years of the wife}
    \item{child5}{number of children for ages 0 to 5}
    \item{child13}{number of children for ages 6 to 13}
    \item{child17}{number of children for ages 14 to 17}
    \item{nonwhite}{non--white ?}
    \item{owned}{is the home owned by the household ?}
    \item{mortgage}{is the home on mortgage ?}
    \item{occupation}{occupation of the husband, one of mp (manager or}
    \item{unemp}{local unemployment rate in \%}
  }
}
\source{
    Lee, Myoung--Jae (1995) \dQuote{Semi--parametric estimation of simultaneous equations with limited dependent variables : a case study of female labour supply}, \emph{Journal of Applied Econometrics}, \bold{10(2)}, april--june, 187--200.

}
\references{
      Journal of Applied Econometrics data archive : \url{http://qed.econ.queensu.ca/jae/}.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}


\keyword{datasets}
