\name{Index.Econometrics}
\alias{Index.Econometrics}
\title{Econometric fields}
\description{
  \describe{
    \itemize{
      \item binomial model
      \itemize{
	\item \code{\link{Benefits}} : Unemployement of Blue Collar Workers
	\item \code{\link{Hdma}} : The Boston HDMA Data Set
	\item \code{\link{Mroz}} : Labor Supply Data
	\item \code{\link{Participation}} : Labor Force Participation
	\item \code{\link{Train}} : Stated Preferences for Train Traveling
      }
      \item censored and truncated model
      \itemize{
	\item \code{\link{Fair}} : Extramarital Affairs Data
	\item \code{\link{HI}} : Health Insurance and Hours Worked By Wives
	\item \code{\link{Mofa}} : International Expansion of U.S. Mofa's (majority--owned Foreign Affiliates in Fire (finance, Insurance and Real Estate)
	\item \code{\link{Tobacco}} : Households Tobacco Budget Share
	\item \code{\link{Workinghours}} : Wife Working Hours
      }
      \item count data
      \itemize{
	\item \code{\link{Accident}} : Ship Accidents
	\item \code{\link{Bids}} : Bids Received By U.S. Firms
	\item \code{\link{Doctor}} : Number of Doctor Visits
	\item \code{\link{DoctorAUS}} : Doctor Visits in Australia
	\item \code{\link{DoctorContacts}} : Contacts With Medical Doctor
	\item \code{\link{OFP}} : Visits to Physician Office
	\item \code{\link{PatentsHGH}} : Dynamic Relation Between Patents and R\&D
	\item \code{\link{PatentsRD}} : Patents, R\&D and Technological Spillovers for a Panel of Firms
	\item \code{\link{Somerville}} : Visits to Lake Somerville
	\item \code{\link{StrikeNb}} : Number of Strikes in Us Manufacturing
      }
      \item duration model
      \itemize{
	\item \code{\link{Oil}} : Oil Investment
	\item \code{\link{Strike}} : Strike Duration Data
	\item \code{\link{StrikeDur}} : Strikes Duration
	\item \code{\link{UnempDur}} : Unemployment Duration
	\item \code{\link{Unemployment}} : Unemployment Duration
      }
      \item multinomial model
      \itemize{
	\item \code{\link{Car}} : Stated Preferences for Car Choice
	\item \code{\link{Catsup}} : Choice of Brand for Catsup
	\item \code{\link{Cracker}} : Choice of Brand for Crakers
	\item \code{\link{Fishing}} : Choice of Fishing Mode
	\item \code{\link{HC}} : Heating and Cooling System Choice in Newly Built Houses in California
	\item \code{\link{Heating}} : Heating System Choice in California Houses
	\item \code{\link{Ketchup}} : Choice of Brand for Ketchup
	\item \code{\link{Mode}} : Mode Choice
	\item \code{\link{ModeChoice}} : Data to Study Travel Mode Choice
	\item \code{\link{Tuna}} : Choice of Brand for Tuna
	\item \code{\link{Yogurt}} : Choice of Brand for Yogurts
      }
      \item ordered model
      \itemize{
	\item \code{\link{Kakadu}} : Willingness to Pay for the Preservation of the Kakadu National Park
	\item \code{\link{Mathlevel}} : Level of Calculus Attained for Students Taking Advanced Micro--economics
	\item \code{\link{NaturalPark}} : Willingness to Pay for the Preservation of the Alentejo Natural Park
      }
      \item panel
      \itemize{
	\item \code{\link{Airline}} : Cost for U.S. Airlines
	\item \code{\link{Cigar}} : Cigarette Consumption
	\item \code{\link{Cigarette}} : The Cigarette Consumption Panel Data Set
	\item \code{\link{Crime}} : Crime in North Carolina
	\item \code{\link{Fatality}} : Drunk Driving Laws and Traffic Deaths
	\item \code{\link{Gasoline}} : Gasoline Consumption
	\item \code{\link{Grunfeld}} : Grunfeld Investment Data
	\item \code{\link{LaborSupply}} : Wages and Hours Worked
	\item \code{\link{Males}} : Wages and Education of Young Males
	\item \code{\link{MunExp}} : Municipal Expenditure Data
	\item \code{\link{Produc}} : Us States Production
	\item \code{\link{SumHes}} : The Penn Table
	\item \code{\link{Wages}} : Panel Datas of Individual Wages
      }
      \item system of equations
      \itemize{
	\item \code{\link{BudgetItaly}} : Budget Shares for Italian Households
	\item \code{\link{BudgetUK}} : Budget Shares of British Households
	\item \code{\link{Electricity}} : Cost Function for Electricity Producers
	\item \code{\link{Klein}} : Klein's Model I
	\item \code{\link{ManufCost}} : Manufacturing Costs
	\item \code{\link{Nerlove}} : Cost Function for Electricity Producers, 1955
	\item \code{\link{University}} : Provision of University Teaching and Research
      }
      \item time--series
      \itemize{
	\item \code{\link{CRSPday}} : Daily Returns from the CRSP Database
	\item \code{\link{CRSPmon}} : Monthly Returns from the CRSP Database
	\item \code{\link{Capm}} : Stock Market Data
	\item \code{\link{Consumption}} : Quarterly Data on Consumption and Expenditure
	\item \code{\link{DM}} : DM Dollar Exchange Rate
	\item \code{\link{Forward}} : Exchange Rates of US Dollar Against Other Currencies
	\item \code{\link{Garch}} : Daily Observations on Exchange Rates of the US Dollar Against Other Currencies
	\item \code{\link{Hstarts}} : Housing Starts
	\item \code{\link{Icecream}} : Ice Cream Consumption
	\item \code{\link{IncomeUK}} : Seasonally Unadjusted Quarterly Data on Disposable Income and Expenditure
	\item \code{\link{Irates}} : Monthly Interest Rates
\item \code{\link{LT}} : Dollar Sterling Exchange Rate
\item \code{\link{MW}} : Growth of Disposable Income and Treasury Bill Rate
\item \code{\link{Macrodat}} : Macroeconomic Time Series for the United States
\item \code{\link{Mishkin}} : Inflation and Interest Rates
\item \code{\link{MoneyUS}} : Macroeconomic Series for the United States
\item \code{\link{Mpyr}} : Money, National Product and Interest Rate
\item \code{\link{Orange}} : The Orange Juice Data Set
\item \code{\link{PE}} : Price and Earnings Index
\item \code{\link{PPP}} : Exchange Rates and Price Indices for France and Italy
\item \code{\link{Pound}} : Pound-dollar Exchange Rate
\item \code{\link{Pricing}} : Returns of Size-based Portfolios
\item \code{\link{Solow}} : Solow's Technological Change Data
\item \code{\link{Tbrate}} : Interest Rate, GDP and Inflation
\item \code{\link{Yen}} : Yen-dollar Exchange Rate
}
}
}
}

\keyword{documentation}
