\name{Index.Economics}
\docType{documentation}
\alias{Index.Economics}
\title{Economic fields}
\description{
  \itemize{
    \item consumer behavior 
    \itemize{
\item \code{\link{BudgetFood}} : Budget Share of Food for Spanish Households  
\item \code{\link{BudgetItaly}} : Budget Shares for Italian Households  
\item \code{\link{BudgetUK}} : Budget Shares of British Households  
\item \code{\link{Car}} : Stated Preferences for Car Choice  
\item \code{\link{Cigar}} : Cigarette Consumption  
\item \code{\link{Cigarette}} : The Cigarette Consumption Panel Data Set  
\item \code{\link{Doctor}} : Number of Doctor Visits  
\item \code{\link{Fishing}} : Choice of Fishing Mode  
\item \code{\link{Gasoline}} : Gasoline Consumption  
\item \code{\link{HC}} : Heating and Cooling System Choice in Newly Built Houses in California  
\item \code{\link{Heating}} : Heating System Choice in California Houses  
\item \code{\link{Icecream}} : Ice Cream Consumption  
\item \code{\link{Mode}} : Mode Choice  
\item \code{\link{ModeChoice}} : Data to Study Travel Mode Choice  
\item \code{\link{Somerville}} : Visits to Lake Somerville  
\item \code{\link{Tobacco}} : Households Tobacco Budget Share  
\item \code{\link{Train}} : Stated Preferences for Train Traveling  
}
\item economics of education 
 \itemize{
\item \code{\link{Caschool}} : The California Test Score Data Set  
\item \code{\link{MCAS}} : The Massashusets Test Score Data Set  
\item \code{\link{Mathlevel}} : Level of Calculus Attained for Students Taking Advanced Micro--economics  
\item \code{\link{Star}} : Effects on Learning of Small Class Sizes  
}
\item environmental economics 
 \itemize{
\item \code{\link{Airq}} : Air Quality for Californian Metropolitan Areas  
\item \code{\link{Kakadu}} : Willingness to Pay for the Preservation of the Kakadu National Park  
\item \code{\link{NaturalPark}} : Willingness to Pay for the Preservation of the Alentejo Natural Park  
}
\item finance 
 \itemize{
\item \code{\link{CRSPday}} : Daily Returns from the CRSP Database  
\item \code{\link{CRSPmon}} : Monthly Returns from the CRSP Database  
\item \code{\link{Capm}} : Stock Market Data  
\item \code{\link{DM}} : DM Dollar Exchange Rate  
\item \code{\link{Forward}} : Exchange Rates of US Dollar Against Other Currencies  
\item \code{\link{Garch}} : Daily Observations on Exchange Rates of the US Dollar Against Other Currencies  
\item \code{\link{Intratesm}} : Interest Rates (monthly)  
\item \code{\link{Intratesq}} : Interest Rates (quarterly)  
\item \code{\link{Irates}} : Monthly Interest Rates  
\item \code{\link{LT}} : Dollar Sterling Exchange Rate  
\item \code{\link{PPP}} : Exchange Rates and Price Indices for France and Italy  
\item \code{\link{Pound}} : Pound-dollar Exchange Rate  
\item \code{\link{Pricing}} : Returns of Size-based Portfolios  
\item \code{\link{Yen}} : Yen-dollar Exchange Rate  
}
\item game theory 
 \itemize{
\item \code{\link{FriendFoe}} : Data from the Television Game Show Friend Or Foe ?  
}
\item health economics 
 \itemize{
\item \code{\link{DoctorAUS}} : Doctor Visits in Australia  
\item \code{\link{DoctorContacts}} : Contacts With Medical Doctor  
\item \code{\link{MedExp}} : Structure of Demand for Medical Care  
\item \code{\link{OFP}} : Visits to Physician Office  
\item \code{\link{VietNamH}} : Medical Expenses in Viet--nam (household Level)  
\item \code{\link{VietNamI}} : Medical Expenses in Viet--nam (individual Level)  
}
\item hedonic prices 
 \itemize{
\item \code{\link{Computers}} : Prices of Personal Computers  
\item \code{\link{Diamond}} : Pricing the C's of Diamond Stones  
\item \code{\link{Hedonic}} : Hedonic Prices of Cencus Tracts in Boston  
\item \code{\link{Housing}} : Sales Prices of Houses in the City of Windsor  
\item \code{\link{Journals}} : Economic Journals Dat Set  
}
\item labor economics 
 \itemize{
\item \code{\link{Benefits}} : Unemployement of Blue Collar Workers  
\item \code{\link{Bwages}} : Wages in Belgium  
\item \code{\link{CPSch3}} : Earnings from the Current Population Survey  
\item \code{\link{Earnings}} : Earnings for Three Age Groups  
\item \code{\link{Griliches}} : Wage Datas  
\item \code{\link{HI}} : Health Insurance and Hours Worked By Wives  
\item \code{\link{LaborSupply}} : Wages and Hours Worked  
\item \code{\link{Labour}} : Belgian Firms  
\item \code{\link{Males}} : Wages and Education of Young Males  
\item \code{\link{Mroz}} : Labor Supply Data  
\item \code{\link{PSID}} : Panel Survey of Income Dynamics  
\item \code{\link{Participation}} : Labor Force Participation  
\item \code{\link{RetSchool}} : Return to Schooling  
\item \code{\link{Schooling}} : Wages and Schooling  
\item \code{\link{Strike}} : Strike Duration Data  
\item \code{\link{StrikeDur}} : Strikes Duration  
\item \code{\link{StrikeNb}} : Number of Strikes in Us Manufacturing  
\item \code{\link{Treatment}} : Evaluating Treatment Effect of Training on Earnings  
\item \code{\link{UnempDur}} : Unemployment Duration  
\item \code{\link{Unemployment}} : Unemployment Duration  
\item \code{\link{Wages}} : Panel Datas of Individual Wages  
\item \code{\link{Wages1}} : Wages, Experience and Schooling  
\item \code{\link{Workinghours}} : Wife Working Hours  
}
\item macroeconomics 
 \itemize{
\item \code{\link{Consumption}} : Quarterly Data on Consumption and Expenditure  
\item \code{\link{Hstarts}} : Housing Starts  
\item \code{\link{IncomeUK}} : Seasonally Unadjusted Quarterly Data on Disposable Income and Expenditure  
\item \code{\link{Klein}} : Klein's Model I  
\item \code{\link{Longley}} : The Longley Data  
\item \code{\link{MW}} : Growth of Disposable Income and Treasury Bill Rate  
\item \code{\link{Macrodat}} : Macroeconomic Time Series for the United States  
\item \code{\link{Mishkin}} : Inflation and Interest Rates  
\item \code{\link{Money}} : Money, GDP and Interest Rate in Canada  
\item \code{\link{MoneyUS}} : Macroeconomic Series for the United States  
\item \code{\link{Mpyr}} : Money, National Product and Interest Rate  
\item \code{\link{PE}} : Price and Earnings Index  
\item \code{\link{Produc}} : Us States Production  
\item \code{\link{Solow}} : Solow's Technological Change Data  
\item \code{\link{SumHes}} : The Penn Table  
\item \code{\link{Tbrate}} : Interest Rate, GDP and Inflation  
}
\item marketing 
 \itemize{
\item \code{\link{Catsup}} : Choice of Brand for Catsup  
\item \code{\link{Cracker}} : Choice of Brand for Crakers  
\item \code{\link{Ketchup}} : Choice of Brand for Ketchup  
\item \code{\link{Tuna}} : Choice of Brand for Tuna  
\item \code{\link{Yogurt}} : Choice of Brand for Yogurts  
}
\item producer behavior 
 \itemize{
\item \code{\link{Accident}} : Ship Accidents  
\item \code{\link{Airline}} : Cost for U.S. Airlines  
\item \code{\link{Bids}} : Bids Received By U.S. Firms  
\item \code{\link{Clothing}} : Sales Data of Men's Fashion Stores  
\item \code{\link{Electricity}} : Cost Function for Electricity Producers  
\item \code{\link{Grunfeld}} : Grunfeld Investment Data  
\item \code{\link{Hdma}} : The Boston HDMA Data Set  
\item \code{\link{ManufCost}} : Manufacturing Costs  
\item \code{\link{Metal}} : Production for SIC 33  
\item \code{\link{Mofa}} : International Expansion of U.S. Mofa's (majority--owned Foreign Affiliates in Fire (finance, Insurance and Real Estate)  
\item \code{\link{Nerlove}} : Cost Function for Electricity Producers, 1955  
\item \code{\link{Oil}} : Oil Investment  
\item \code{\link{Orange}} : The Orange Juice Data Set  
\item \code{\link{PatentsHGH}} : Dynamic Relation Between Patents and R\&D  
\item \code{\link{PatentsRD}} : Patents, R\&D and Technological Spillovers for a Panel of Firms  
\item \code{\link{TranspEq}} : Statewide Data on Transportation Equipment Manufacturing  
\item \code{\link{University}} : Provision of University Teaching and Research  
}
\item socio--economics 
 \itemize{
\item \code{\link{Crime}} : Crime in North Carolina  
\item \code{\link{Fair}} : Extramarital Affairs Data  
\item \code{\link{Fatality}} : Drunk Driving Laws and Traffic Deaths  
}
}
}



\keyword{documentation}
