\name{Index.Observations}
\alias{Index.Observations}
\title{Observations}
\description{
  \describe{
    \itemize{
      \item country
      \itemize{
        \item \code{\link{Consumption}} : Quarterly 
          Data on Consumption and Expenditure
        \item \code{\link{DM}} : DM Dollar Exchange
          Rate
        \item \code{\link{Garch}} : Daily
          Observations on Exchange Rates of the US
          Dollar Against Other Currencies
        \item \code{\link{Gasoline}} : Gasoline
          Consumption
        \item \code{\link{Hstarts}} : Housing 
          Starts
        \item \code{\link{Icecream}} : Ice Cream
          Consumption
        \item \code{\link{IncomeUK}} : Seasonally
          Unadjusted Quarterly Data on Disposable
          Income and Expenditure
        \item \code{\link{Irates}} : Monthly 
          Interest Rates
        \item \code{\link{Klein}} : Klein's Model I
        \item \code{\link{LT}} : Dollar Sterling
          Exchange Rate
        \item \code{\link{Longley}} : The Longley
          Data
        \item \code{\link{MW}} : Growth of 
          Disposable Income and Treasury Bill Rate
        \item \code{\link{Macrodat}} : 
          Macroeconomic Time Series for the United
          States
        \item \code{\link{ManufCost}} : 
          Manufacturing Costs
        \item \code{\link{Mishkin}} : Inflation 
          and Interest Rates
        \item \code{\link{Mofa}} : International
          Expansion of U.S. MOFAs (majority--owned
          Foreign Affiliates in Fire (finance,
          Insurance and Real Estate)
        \item \code{\link{Money}} : Money, GDP and
          Interest Rate in Canada
        \item \code{\link{Mpyr}} : Money, National
          Product and Interest Rate
        \item \code{\link{Orange}} : The Orange 
          Juice Data Set
        \item \code{\link{PE}} : Price and 
          Earnings Index
        \item \code{\link{PPP}} : Exchange Rates 
          and Price Indices for France and Italy
        \item \code{\link{Pound}} : Pound-dollar
          Exchange Rate
        \item \code{\link{Solow}} : Solow's
          Technological Change Data
        \item \code{\link{StrikeNb}} : Number of
          Strikes in Us Manufacturing
        \item \code{\link{SumHes}} : The Penn Table
        \item \code{\link{Tbrate}} : Interest Rate,
          GDP and Inflation
        \item \code{\link{Yen}} : Yen-dollar 
          Exchange Rate
}
\item goods
 \itemize{
    \item \code{\link{Computers}} : Prices of
      Personal Computers
    \item \code{\link{Diamond}} : Pricing the C's 
      of Diamond Stones
    \item \code{\link{Housing}} : Sales Prices of
      Houses in the City of Windsor
    \item \code{\link{Journals}} : Economic Journals
      Data Set
}
\item households
 \itemize{
\item \code{\link{BudgetFood}} : Budget Share of Food for Spanish Households
\item \code{\link{BudgetItaly}} : Budget Shares for Italian Households
\item \code{\link{BudgetUK}} : Budget Shares of British Households
\item \code{\link{HC}} : Heating and Cooling System Choice in Newly Built Houses in California
\item \code{\link{Heating}} : Heating System Choice in California Houses
\item \code{\link{VietNamH}} : Medical Expenses in Vietnam (household Level)
}
\item individuals
 \itemize{
\item \code{\link{Benefits}} : Unemployment of Blue Collar Workers
\item \code{\link{Bwages}} : Wages in Belgium
\item \code{\link{CPSch3}} : Earnings from the Current Population Survey
\item \code{\link{Car}} : Stated Preferences for Car Choice
\item \code{\link{Catsup}} : Choice of Brand for Catsup
\item \code{\link{Cracker}} : Choice of Brand for Crackers
\item \code{\link{Doctor}} : Number of Doctor Visits
\item \code{\link{DoctorAUS}} : Doctor Visits in Australia
\item \code{\link{Earnings}} : Earnings for Three Age Groups
\item \code{\link{Fair}} : Extramarital Affairs Data
\item \code{\link{Fishing}} : Choice of Fishing Mode
\item \code{\link{FriendFoe}} : Data from the Television Game Show Friend Or Foe ?
\item \code{\link{Griliches}} : Wage Data
\item \code{\link{HI}} : Health Insurance and Hours Worked By Wives
\item \code{\link{Hmda}} : The Boston HMDA Data Set
\item \code{\link{Kakadu}} : Willingness to Pay for the Preservation of the Kakadu National Park
\item \code{\link{Ketchup}} : Choice of Brand for Ketchup
\item \code{\link{Males}} : Wages and Education of Young Males
\item \code{\link{Mathlevel}} : Level of Calculus Attained for Students Taking Advanced Micro--economics
\item \code{\link{Mode}} : Mode Choice
\item \code{\link{ModeChoice}} : Data to Study Travel Mode Choice
\item \code{\link{Mroz}} : Labor Supply Data
\item \code{\link{NaturalPark}} : Willingness to Pay for the Preservation of the Alentejo Natural Park
\item \code{\link{OFP}} : Visits to Physician Office
\item \code{\link{PSID}} : Panel Survey of Income Dynamics
\item \code{\link{Participation}} : Labor Force Participation
\item \code{\link{RetSchool}} : Return to Schooling
\item \code{\link{Schooling}} : Wages and Schooling
\item \code{\link{Somerville}} : Visits to Lake Somerville
\item \code{\link{Star}} : Effects on Learning of Small Class Sizes
\item \code{\link{Tobacco}} : Households Tobacco Budget Share
\item \code{\link{Train}} : Stated Preferences for Train Traveling
\item \code{\link{Tuna}} : Choice of Brand for Tuna
\item \code{\link{Unemployment}} : Unemployment Duration
\item \code{\link{VietNamI}} : Medical Expenses in Vietnam (individual Level)
\item \code{\link{Wages}} : Panel Data of Individual Wages
\item \code{\link{Wages1}} : Wages, Experience and Schooling
\item \code{\link{Workinghours}} : Wife Working Hours
\item \code{\link{Yogurt}} : Choice of Brand for Yogurts
}
\item production units
 \itemize{
\item \code{\link{Airline}} : Cost for U.S. Airlines
\item \code{\link{Bids}} : Bids Received By U.S. Firms
\item \code{\link{CRSPday}} : Daily Returns from the CRSP Database
\item \code{\link{CRSPmon}} : Monthly Returns from the CRSP Database
\item \code{\link{Clothing}} : Sales Data of Men's Fashion Stores
\item \code{\link{Electricity}} : Cost Function for Electricity Producers
\item \code{\link{Grunfeld}} : Grunfeld Investment Data
\item \code{\link{Labour}} : Belgian Firms
\item \code{\link{Nerlove}} : Cost Function for Electricity Producers, 1955
\item \code{\link{Oil}} : Oil Investment
\item \code{\link{PatentsHGH}} : Dynamic Relation Between Patents and R\&D
\item \code{\link{PatentsRD}} : Patents, R\&D and Technological Spillovers for a Panel of Firms
}
\item regional
 \itemize{
\item \code{\link{Airq}} : Air Quality for Californian Metropolitan Areas
\item \code{\link{Cigar}} : Cigarette Consumption
\item \code{\link{Cigarette}} : The Cigarette Consumption Panel Data Set
\item \code{\link{Crime}} : Crime in North Carolina
\item \code{\link{Fatality}} : Drunk Driving Laws and Traffic Deaths
\item \code{\link{Hedonic}} : Hedonic Prices of Census Tracts in Boston
\item \code{\link{Metal}} : Production for SIC 33
\item \code{\link{MunExp}} : Municipal Expenditure Data
\item \code{\link{Produc}} : Us States Production
\item \code{\link{TranspEq}} : Statewide Data on Transportation Equipment Manufacturing
}
\item schools
 \itemize{
\item \code{\link{Caschool}} : The California Test Score Data Set
\item \code{\link{MCAS}} : The Massachusetts Test Score Data Set
\item \code{\link{University}} : Provision of University Teaching and Research
}
}
}

}
\keyword{documentation}
