\name{Index.Time.Series}
\docType{documentation}
\alias{Index.Time.Series}
\title{Time Series}
\description{
  \describe{
    \itemize{
\item annual 
 \itemize{
\item \code{\link{Klein}} : Klein's Model I  
\item \code{\link{LT}} : Dollar Sterling Exchange Rate  
\item \code{\link{Longley}} : The Longley Data  
\item \code{\link{ManufCost}} : Manufacturing Costs  
\item \code{\link{Mpyr}} : Money, National Product and Interest Rate  
\item \code{\link{PE}} : Price and Earnings Index  
\item \code{\link{Solow}} : Solow's Technological Change Data  
}
\item daily 
 \itemize{
\item \code{\link{CRSPday}} : Daily Returns from the CRSP Database  
\item \code{\link{Garch}} : Daily Observations on Exchange Rates of the US Dollar Against Other Currencies  
\item \code{\link{SP500}} : Returns on Standard \& Poor's 500 Index  
}
\item four--weekly 
 \itemize{
\item \code{\link{Icecream}} : Ice Cream Consumption  
}
\item monthly 
 \itemize{
\item \code{\link{CRSPmon}} : Monthly Returns from the CRSP Database  
\item \code{\link{Capm}} : Stock Market Data  
\item \code{\link{Forward}} : Exchange Rates of US Dollar Against Other Currencies  
\item \code{\link{Intratesm}} : Interest Rates (monthly)  
\item \code{\link{Irates}} : Monthly Interest Rates  
\item \code{\link{Mishkin}} : Inflation and Interest Rates  
\item \code{\link{Orange}} : The Orange Juice Data Set  
\item \code{\link{PPP}} : Exchange Rates and Price Indices for France and Italy  
\item \code{\link{Pricing}} : Returns of Size-based Portfolios  
\item \code{\link{StrikeNb}} : Number of Strikes in Us Manufacturing  
}
\item quarterly 
 \itemize{
\item \code{\link{Consumption}} : Quarterly Data on Consumption and Expenditure  
\item \code{\link{Hstarts}} : Housing Starts  
\item \code{\link{IncomeUK}} : Seasonally Unadjusted Quarterly Data on Disposable Income and Expenditure  
\item \code{\link{Intratesq}} : Interest Rates (quarterly)  
\item \code{\link{MW}} : Growth of Disposable Income and Treasury Bill Rate  
\item \code{\link{Macrodat}} : Macroeconomic Time Series for the United States  
\item \code{\link{Money}} : Money, GDP and Interest Rate in Canada  
\item \code{\link{MoneyUS}} : Macroeconomic Series for the United States  
\item \code{\link{Tbrate}} : Interest Rate, GDP and Inflation  
}
\item weekly 
 \itemize{
\item \code{\link{DM}} : DM Dollar Exchange Rate  
\item \code{\link{Pound}} : Pound-dollar Exchange Rate  
\item \code{\link{Yen}} : Yen-dollar Exchange Rate  
}
}
}
}


\keyword{documentation}
