\name{Macrodat}
\docType{data}
\alias{Macrodat}
\title{Macroeconomic Time Series for the United States }
\description{
  quarterly observations from 1959-1 to 2000-4

 \emph{number of observations} :  168

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(Macrodat)}
\format{A time serie containing :
  \describe{
    \item{lhur}{unemployment Rate (Average of Months in Quarter)       }
    \item{punew}{CPI   (Average of Months in Quarter)       }
    \item{fyff}{federal Funds Interest Rate (Last Month in Quarter)       }
    \item{fygm3}{3 Month Treasury Bill Interest Rate (Last Month in Quarter)       }
    \item{fygt1}{1 Year Treasury Bond Interest Rate (Last Month in Quarter)       }
    \item{exruk}{dollar / Pound Exchange Rate (Last Month in Quarter)       }
    \item{gdpjp}{real GDP for Japan       }
  }
}
\source{
    Bureau of Labor Statistics, OECD, Federal Reserve.

}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers, \url{http://wps.aw.com/aw_stockwatsn_economtrcs_1}.
}
\seealso{\code{\link{Index.Books}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}, \code{\link{Index.Time.Series}}}

\keyword{datasets}
