\name{PE}
\docType{data}
\alias{PE}
\title{Price and Earnings Index }
\description{
  annual observations from 1800 to 1931

 \emph{number of observations} :  132

 \emph{observation} :  country

   \emph{country} :  United States

}
\usage{data(PE)}
\format{A time series containing :
  \describe{
    \item{price}{S\&P composite stock price index}
    \item{earnings}{S\&P composite earnings index}
  }
}
\source{
    Robert Shiller.

}
\references{
    Verbeek, Marno (2004) \emph{A Guide to Modern Econometrics}, John Wiley and Sons, chapter 8.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
