\name{PPP}
\docType{data}
\alias{PPP}
\title{Exchange Rates and Price Indices for France and Italy }
\description{
  monthly observations from 1981--01 to 1996--06

 \emph{number of observations} :  186

 \emph{observation} :  country

   \emph{country} :  France and Italy

}
\usage{data(PPP)}
\format{A time series containing :
  \describe{
    \item{lnit}{log price index Italy}
    \item{lnfr}{log price index France}
    \item{lnx}{ log exchange rate France/Italy}
    \item{cpiit}{consumer price index Italy}
    \item{cpifr}{consumer price index France}
  }
}
\source{
    \code{Datastream}.
}
\references{
    Verbeek, Marno (2004) \emph{A Guide to Modern Econometrics}, John Wiley and Sons, chapters 8 and 9.

}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
