\name{SP500}
\docType{data}
\alias{SP500}
\title{Returns on Standard & Poor's 500 Index }
\description{
  daily observations from 1981--01 to 1991--04

 \emph{number of observations} :  2783

}
\usage{data(SP500)}
\format{A dataframe containing :
  \describe{
    \item{r500}{daily return S&P500 (change in log index)}
  }
}
\references{
    Verbeek, Marno (2004) \emph{A Guide to Modern Econometrics}, John Wiley and Sons.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
