\name{Star}
\docType{data}
\alias{Star}
\title{Effects on Learning of Small Class Sizes }
\description{
  a cross-section from 1985-89

 \emph{number of observations} :  5748

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Star)}
\format{A dataframe containing :
  \describe{
    \item{tmathssk}{total math scaled score}
    \item{treadssk}{total reading scaled score}
    \item{classk}{type of class, a factor with levels (regular,small.class,regular.with.aide)}
    \item{totexpk}{years of total teaching experience}
    \item{sexe}{a factor with levels (boy,girl)}
    \item{freelunk}{qualified for free lunch ? a factor}
    \item{race}{a factor with levels (white,black,other)}
    \item{schidkn}{school indicator variable}
  }
}
\source{
    Project STAR \url{http://www.heros-inc.org/star.htm}.

}
\references{
    Stock, James H.  and  Mark W.  Watson (2003) \emph{Introduction to Econometrics}, Addison-Wesley Educational Publishers, \url{http://wps.aw.com/aw_stockwatsn_economtrcs_1}.
}
\seealso{\code{\link{Index.Books}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}}}

\keyword{datasets}
