\name{Wages1}
\docType{data}
\alias{Wages1}
\title{Wages, Experience and Schooling }
\description{
  a panel of 595 observations  from 1976  to 1982

 \emph{number of observations} :  3294

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Wages1)}
\format{A time series containing :
  \describe{
    \item{exper}{experience in years}
    \item{sex}{a factor with levels (male,female)}
    \item{school}{years of schooling}
    \item{wage}{wage (in 1980 $) per hour}
  }
}
\references{
    Verbeek, Marno (2004) \emph{A Guide to Modern Econometrics}, John Wiley and Sons.
}
\seealso{\code{\link{Index.Source}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}, \code{\link{Index.Observations}},

 \code{\link{Index.Time.Series}}}


\keyword{datasets}
